% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{ACI}
\alias{ACI}
\title{123. Atmospheric Clarity Index
ratio between solar radiation at surface and solar radiation at TOA (alt top of the atmosphere empirically obtained, see https://goo.gl/Wzs1Zk)
http://www.greenrhinoenergy.com/solar/radiation/atmosphere.php}
\usage{
ACI(data, toa, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{net radiation}

\item{toa}{solar radiation at TOA}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
ACI
}
\description{
123. Atmospheric Clarity Index
ratio between solar radiation at surface and solar radiation at TOA (alt top of the atmosphere empirically obtained, see https://goo.gl/Wzs1Zk)
http://www.greenrhinoenergy.com/solar/radiation/atmosphere.php
}
\examples{
data(radiation); data(toa)
ACI(data = radiation.value, toa=toa.value)
}
