% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis.R
\name{allow_scale}
\alias{allow_scale}
\title{Function allow scale}
\usage{
allow_scale(name)
}
\arguments{
\item{name}{function name}
}
\value{
allow or not allow
}
\description{
Function allow scale
}
\keyword{internal}
