% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{asd}
\alias{asd}
\title{Average snow depth}
\usage{
asd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
average snow depth
}
\description{
Average snow depth
}
\examples{
data(data_all)
asd(data = data_all$snowdepththickness)
}
