% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{AT}
\alias{AT}
\title{88. Apparent Temperature
AT = Ta + 0.33e -0.70v -4.00; Ta = air temperature in ºC ; v = wind speed in m/s; e= water vapour pressure in hPa}
\usage{
AT(taverage, w, vapor, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{w}{average wind}

\item{vapor}{water vapour pressure}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
AT
}
\description{
88. Apparent Temperature
AT = Ta + 0.33e -0.70v -4.00; Ta = air temperature in ºC ; v = wind speed in m/s; e= water vapour pressure in hPa
}
\examples{
calculate88(taverage = taverage.value, w = w.value, vapor = vapor.value)
}
