% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{average_snow_depth}
\alias{average_snow_depth}
\title{114. Average snow depth during the reference period (annual, seasonal and monthly)}
\usage{
average_snow_depth(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
average snow depth
}
\description{
114. Average snow depth during the reference period (annual, seasonal and monthly)
}
\examples{
calculate114(data = snow_depth.value)
}
