% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{average_temp}
\alias{average_temp}
\title{1. Average temperature (maximum, minimum and mean)}
\usage{
average_temp(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum, minimum or medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
average temperature
}
\description{
1. Average temperature (maximum, minimum and mean)
}
\examples{
calculate1(data=tmax.value)
}
