% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO10}
\alias{BIO10}
\title{69. BIO10
Mean Temperature of Warmest Quarter}
\usage{
BIO10(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO10
}
\description{
69. BIO10
Mean Temperature of Warmest Quarter
}
\examples{
calculate69(data = taverage.value)
}
