% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO11}
\alias{BIO11}
\title{70. BIO11
Mean Temperature of Coldest Quarter}
\usage{
BIO11(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO11
}
\description{
70. BIO11
Mean Temperature of Coldest Quarter
}
\examples{
calculate70(data = taverage.value)
}
