% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO13}
\alias{BIO13}
\title{71. BIO13
Precipitation of Wettest Month}
\usage{
BIO13(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO13
}
\description{
71. BIO13
Precipitation of Wettest Month
}
\examples{
calculate71(data = pr.value)
}
