% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO14}
\alias{BIO14}
\title{72. BIO14
Precipitation of Driest Month}
\usage{
BIO14(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO14
}
\description{
72. BIO14
Precipitation of Driest Month
}
\examples{
calculate72(data = pr.value)
}
