% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO15}
\alias{BIO15}
\title{73. BIO15
Precipitation Seasonality (Coefficient of Variation)}
\usage{
BIO15(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO15
}
\description{
73. BIO15
Precipitation Seasonality (Coefficient of Variation)
}
\examples{
calculate73(data = pr.value)
}
