% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO16}
\alias{BIO16}
\title{74. BIO16
Precipitation of Wettest Quarter}
\usage{
BIO16(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO16
}
\description{
74. BIO16
Precipitation of Wettest Quarter
}
\examples{
calculate74(data = pr.value)
}
