% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO17}
\alias{BIO17}
\title{75. BIO17
Precipitation of Driest Quarter}
\usage{
BIO17(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO17
}
\description{
75. BIO17
Precipitation of Driest Quarter
}
\examples{
calculate75(data = pr.value)
}
