% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO18}
\alias{BIO18}
\title{76. BIO18
Precipitation of Warmest Quarter}
\usage{
BIO18(pr, taverage, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{pr}{precipitation}

\item{taverage}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO18
}
\description{
76. BIO18
Precipitation of Warmest Quarter
}
\examples{
calculate76(pr=pr.value, taverage=taverage.value)
}
