% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO19}
\alias{BIO19}
\title{77. BIO19
Precipitation of Coldest Quarter}
\usage{
BIO19(pr, taverage, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{pr}{precipitation}

\item{taverage}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO19
}
\description{
77. BIO19
Precipitation of Coldest Quarter
}
\examples{
calculate77(pr=pr.value, taverage=taverage.value)
}
