% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO20}
\alias{BIO20}
\title{84. BIO20
UTCI (Universal Thermal Climate Index)  (Blazejczyk et all, 2012) (Air temperature, Humidity, Wind) fully described at https://goo.gl/by4hH9 http://www.utci.org}
\usage{
BIO20(taverage, pr, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO20
}
\description{
84. BIO20
UTCI (Universal Thermal Climate Index)  (Blazejczyk et all, 2012) (Air temperature, Humidity, Wind) fully described at https://goo.gl/by4hH9 http://www.utci.org
}
\examples{
calculate84(taverage = taverage.value, pr = pr.value)
}
