% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO4}
\alias{BIO4}
\title{78. BIO4
Temperature Seasonality (standard deviation *100)}
\usage{
BIO4(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO4
}
\description{
78. BIO4
Temperature Seasonality (standard deviation *100)
}
\examples{
calculate78(data = taverage.value)
}
