% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO5}
\alias{BIO5}
\title{79. BIO5
max Temperature of Warmest Month}
\usage{
BIO5(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
BIO5
}
\description{
79. BIO5
max Temperature of Warmest Month
}
\examples{
calculate79(data = taverage.value)
}
