% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO6}
\alias{BIO6}
\title{80. BIO6
min Temperature of Coldest Month}
\usage{
BIO6(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
BIO6
}
\description{
80. BIO6
min Temperature of Coldest Month
}
\examples{
calculate80(data = taverage.value)
}
