% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO7}
\alias{BIO7}
\title{81. BIO7
Temperature Annual Range (BIO5-BIO6)}
\usage{
BIO7(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
BIO7
}
\description{
81. BIO7
Temperature Annual Range (BIO5-BIO6)
}
\examples{
calculate81(data = taverage.value)
}
