% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO8}
\alias{BIO8}
\title{82. BIO8
Mean Temperature of Wettest Quarter}
\usage{
BIO8(taverage, pr, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO8
}
\description{
82. BIO8
Mean Temperature of Wettest Quarter
}
\examples{
calculate82(taverage = taverage.value, pr = pr.value)
}
