% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{BIO9}
\alias{BIO9}
\title{83. BIO9
Mean Temperature of Driest Quarter}
\usage{
BIO9(taverage, pr, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
BIO9
}
\description{
83. BIO9
Mean Temperature of Driest Quarter
}
\examples{
calculate83(taverage = taverage.value, pr = pr.value)
}
