% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{budyko_index}
\alias{budyko_index}
\title{102. Budyko Index
(PP/LP)*100, where R is the mean annual net radiation (also known as the net radiation balance), PP is the mean annual precipitation, and L is the latent heat of vaporization for water
https://es.wikipedia.org/wiki/Clasificaci%C3%B3n_clim%C3%A1tica_de_Budyko
de vaporizaci처n: 2257 kJ/kg (539,4 cal/g) a 97 째C.}
\usage{
budyko_index(data, pr, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{net radiation}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Budyko Index
}
\description{
102. Budyko Index
(PP/LP)*100, where R is the mean annual net radiation (also known as the net radiation balance), PP is the mean annual precipitation, and L is the latent heat of vaporization for water
https://es.wikipedia.org/wiki/Clasificaci%C3%B3n_clim%C3%A1tica_de_Budyko
de vaporizaci처n: 2257 kJ/kg (539,4 cal/g) a 97 째C.
}
\examples{
calculate102(data = radiation.value, pr = pr.value)
}
