% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{byMonths_chron}
\alias{byMonths_chron}
\title{Data with months and years in names}
\usage{
byMonths_chron(data)
}
\arguments{
\item{data}{data}
}
\value{
dates
}
\description{
Data with months and years in names
}
\examples{
byMonths_chron(data=tmax.value)
}
