% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{calc_eto}
\alias{calc_eto}
\title{Et0}
\usage{
calc_eto(tmin, tmax, insolation, w, lat, tdew, rh = NA)
}
\arguments{
\item{tmin}{minimum temperature}

\item{tmax}{maximum temperature}

\item{insolation}{sunshine duration}

\item{w}{average wind}

\item{lat}{latitude}

\item{tdew}{dew point}

\item{rh}{relative humidity}
}
\value{
et0
}
\description{
Et0
}
\examples{
calc_eto(tmin = tmin.value, tmax = tmax.value, 
insolation = insolation.value, w = w.value, lat=lat, tdew = dew_point.value)
}
