% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{calc_spei}
\alias{calc_spei}
\title{125. SPEI: Standardized Precipitation Evapotranspiration Index
1, 3, 6 and 12 month SPEI}
\usage{
calc_spei(eto, pr, data_names = NULL, scale = 3, na.rm = FALSE)
}
\arguments{
\item{eto}{et0}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{scale}{scale}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
SPEI
}
\description{
125. SPEI: Standardized Precipitation Evapotranspiration Index
1, 3, 6 and 12 month SPEI
}
\keyword{internal}
