% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{calc_spi}
\alias{calc_spi}
\title{SPI: Standardized Precipitation Index
1, 3, 6 and 12 month SPI}
\usage{
calc_spi(data, data_names = NULL, scale = 3, na.rm = FALSE)
}
\arguments{
\item{data}{daily precipitation, mm}

\item{data_names}{names of each period of time}

\item{scale}{scale}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
SPI
}
\description{
SPI: Standardized Precipitation Index
1, 3, 6 and 12 month SPI
}
\keyword{internal}
