% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{calcf_data}
\alias{calcf_data}
\title{Operation de data, agrupando los datos por valores de names}
\usage{
calcf_data(data, date, extract_names, data_names, operation, ...)
}
\arguments{
\item{data}{data}

\item{date}{date}

\item{extract_names}{Operation to split data}

\item{data_names}{names of each period of time}

\item{operation}{Main operation}

\item{...}{...}
}
\value{
result operation
}
\description{
Operation de data, agrupando los datos por valores de names
}
\examples{
calcf_data(data=data, extract_names=years, operation=meanf)
}
