% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{calcf_data_}
\alias{calcf_data_}
\title{Operation de data, agrupando los datos por valores de names}
\usage{
calcf_data_(data_names, data, operation, ...)
}
\arguments{
\item{data_names}{names}

\item{data}{data}

\item{operation}{operation}

\item{...}{...}
}
\value{
operation
}
\description{
Operation de data, agrupando los datos por valores de names
}
\examples{
calc = calcf_data_(names=months_years(chron(names(pr.value))), data=pr.value, operation=maxf)
}
