% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{calcf_data__}
\alias{calcf_data__}
\title{Operation de data para los valores de oks == ok}
\usage{
calcf_data__(ok, oks, data, operation, ...)
}
\arguments{
\item{ok}{ok}

\item{oks}{oks}

\item{data}{data}

\item{operation}{operation}

\item{...}{...}
}
\value{
operation
}
\description{
Operation de data para los valores de oks == ok
}
\examples{
calcf_data__(years[12], years, max, maxf)
}
