% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{calcf_data_max_min_mean}
\alias{calcf_data_max_min_mean}
\title{Estaci칩n del a침o de una fecha dada}
\usage{
calcf_data_max_min_mean(max, min, mean, extract_names, operation, ...)
}
\arguments{
\item{max}{maximum temperature}

\item{min}{minimum temperature}

\item{mean}{medium temperature}

\item{extract_names}{Operation to split data}

\item{operation}{Main operation}

\item{...}{...}
}
\value{
result operation
}
\description{
Estaci칩n del a침o de una fecha dada
}
\examples{
calcf_data_max_min_mean(time)
}
