% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{calculate_all}
\alias{calculate_all}
\title{Calculate all indexes}
\usage{
calculate_all(tmin.value = NULL, tmax.value = NULL, taverage.value = NULL,
  insolation.value = NULL, w.value = NULL, w_max.value = NULL,
  pr.value = NULL, snow.value = NULL, snow_depth.value = NULL,
  cloud_cover.value = NULL, cloud_cover_less_100.value = NULL,
  radiation.value = NULL, dew_point.value = NULL, toa.value = NULL,
  lat = NULL, time.scale = YEAR, data_names = NULL)
}
\arguments{
\item{tmin.value}{minimum temperature}

\item{tmax.value}{maximum temperature}

\item{taverage.value}{medium temperature}

\item{insolation.value}{sunshine duration}

\item{w.value}{average wind}

\item{w_max.value}{maximum wind gust}

\item{pr.value}{precipitation}

\item{snow.value}{snowfall}

\item{snow_depth.value}{snow depth}

\item{cloud_cover.value}{cloud cover}

\item{cloud_cover_less_100.value}{cloud base below 100 meter}

\item{radiation.value}{net radiation}

\item{dew_point.value}{dew point}

\item{toa.value}{toa}

\item{lat}{latitude}

\item{time.scale}{month, season or year}

\item{data_names}{names of each period of time}
}
\value{
all indexes
}
\description{
Calculate all indexes
}
\examples{
calculate_all(tmin.value=tmin.value, tmax.value=tmax.value,
taverage.value=taverage.value, insolation.value=insolation.value,
w.value=w.value, w_max.value=w_max.value, pr.value=pr.value,
snow.value=snow.value, snow_depth.value=snow_depth.value,
cloud_cover.value=cloud_cover.value,
cloud_cover_less_100.value=cloud_cover_less_100.value,
radiation.value=radiation.value, dew_point.value=dew_point.value,
lat=lat)
}
