% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{cc}
\alias{cc}
\title{Mean daily cloud cover}
\usage{
cc(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{cloud cover}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
mean CC
}
\description{
Mean daily cloud cover (%)
}
\examples{
data(data_all)
cc(data = data_all$cloud)
}
