% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{cdd}
\alias{cdd}
\title{Cold spell duration}
\usage{
cdd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Cold spell duration index
}
\description{
Count of days with at least 6 consecutive days when TN < 10th percentile
}
\examples{
data(data_all)
cdd(data=data_all$tn)
}
