% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{cfd}
\alias{cfd}
\title{Maximum consecutive frost days}
\usage{
cfd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
maximum consecutive frost
}
\description{
Maximum number of consecutive days with TN < 0 Celsius
}
\examples{
data(data_all)
cfd(data=data_all$tn)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
