% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.cc}
\alias{climdex.cc}
\title{Mean cloud cover}
\usage{
climdex.cc(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean cloud cover in octa or percent)}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the time series of mean cloud cover.
}
\description{
This function computes the climdex index CC

This function takes a climdexInput object as input and computes the 
CLOUD_MEAN index: The mean cloud cover measured within a period.
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
