% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.cc2}
\alias{climdex.cc2}
\title{Mostly sunny days}
\usage{
climdex.cc2(ci, freq = c("annual", "halfyear", "seasonal", "monthly"),
  unit = "octa")
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean cloud cover in octa or percent)}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}

\item{unit}{unit of the cloud cover. Allowed are: "octa" or "percent". Default: "octa".}
}
\value{
A vector containing the number of mostly sunny days.
}
\description{
This function computes the climdex index CC2: The number of mostly sunny days (Cloud cover <= 2octa / 20%) within a period.
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
