% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.cd}
\alias{climdex.cd}
\title{Compound Indices}
\usage{
climdex.cd(ci, freq = c("monthly", "annual", "halfyear", "seasonal"),
  precip.thresh = "q25", precip.op = "<", temp.thresh = "q25",
  temp.op = "<")
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily precipitation [mm] and the averaged daily temperature [C])}
}
\value{
A vector containing an annual timeseries of precipitation in wet days.
}
\description{
This function takes a climdexInput object as input and computes the climdex
index CD: the number of days where TG<25 & RR<25 (for wet days: days where precipitation is at least 1mm).
Note: this function doesnt use directly tavg but derives it from tmax & tmin (concistent with .ncdf)
}
\details{
Cold-dry days

This function computes the climdex index CD.
}
