% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.cdd}
\alias{climdex.cdd}
\title{Maximum Consecutive Dry Days}
\usage{
climdex.cdd(ci, freq = c("monthly", "annual", "halfyear", "seasonal"),
  spells.can.span.years = TRUE)
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{spells.can.span.years}{Whether to allow spells to span years}
}
\value{
A vector containing the length of the spell for each year.
}
\description{
This function computes the climdex index CDD.

The \code{spells.can.span.years} option specifies whether spells can cross
year boundaries -- i.e., span years. The default for this is the same as for
fclimdex.
}
\details{
This function computes the climdex index CDD: the annual maximum length of dry spells, in days.
Dry spells are considered to be sequences of days where daily preciptation
is less than 1mm per day.
}
\references{
\url{http://etccdi.pacificclimate.org/list_27_indices.shtml}
}
\seealso{
\code{\link{climdexInput.raw}}, \code{\link{climdexInput.csv}},
\code{\link{spell.length.max}}.
}
\keyword{climate}
\keyword{ts}
