% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{climdex.csu}
\alias{climdex.csu}
\title{Consecutive Summer Days}
\usage{
climdex.csu(ci, freq = c("monthly", "annual"),
  spells.can.span.years = FALSE)
}
\arguments{
\item{ci}{Object of type climdexInput. Here the daily maximum temperature.}

\item{freq}{Time frequency to aggregate to. Allowed only monthly, annual, halfyear, seasonal.}

\item{spells.can.span.years}{Default FALSE}
}
\value{
A vector containing a timeseries of the number of consecutive summer days in a given period (freq).
}
\description{
This function takes a climdexInput object as input and computes the climdex
index csu: the annual (or at different periods) count of consecutive summer days (TX >25C)
}
\author{
Christiana Photiadou (KNMI)
}
\references{
\url{http://www.ecad.eu/documents/atbd.pdf}
}

