% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{climdex.cw}
\alias{climdex.cw}
\title{Cold-wet days}
\usage{
climdex.cw(ci, freq = c("annual", "monthly", "halfyear", "seasonal"),
  precip.thresh = "q75", precip.op = ">", temp.thresh = "q25",
  temp.op = "<")
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily precipitation [mm] and the averaged daily temperature [C])}
}
\value{
A vector containing an annual timeseries of precipitation in wet days.
}
\description{
This function computes the climdex index CW.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index CW: the number of days where TG<25 & RR>75 (for wet days: days where precipitation is at least 1mm).
}
\references{
\url{http://etccdi.pacificclimate.org/list_27_indices.shtml}
}
\seealso{
\code{\link{climdexInput.raw}}, \code{\link{climdexInput.csv}}.
}
\keyword{climate}
\keyword{ts}
