% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.ddnorth}
\alias{climdex.ddnorth}
\title{Northerly winds}
\usage{
climdex.ddnorth(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean wind direction in [deg] with 0deg being wind from north)}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the time series of days with northerly winds.
}
\description{
This function computes the climdex index DDnorth: Days with northerly wind (-45deg (315deg) < wind_dir <= 45deg).
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
