% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.ddwest}
\alias{climdex.ddwest}
\title{Westerly winds}
\usage{
climdex.ddwest(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean wind direction in [deg] with 270deg being wind from west)}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the time series of days with westerly winds.
}
\description{
This function computes the climdex index DDwest: Days with westerly wind (225deg < wind_dir <= 315deg).
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
