% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.dtr}
\alias{climdex.dtr}
\title{Mean Diurnal Temperature Range}
\usage{
climdex.dtr(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}
}
\value{
A vector containing the mean monthly or mean annual diurnal
temperature range.
}
\description{
This function computes the diurnal temperature range on a monthly basis.
}
\details{
\code{climdex.dtr} computes the mean daily diurnal temperature range. The
frequency of observation can be either monthly or annual.
}
\note{
This function creates results which may differ in the 3rd decimal
place from the results from fclimdex.
}
