% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.fd}
\alias{climdex.fd}
\title{Frost Days}
\usage{
climdex.fd(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the number of frost days for each year.
}
\description{
This function computes the climdex index FD.

This function takes a climdexInput object as input and computes the FD (frost
days) climdex index: that is, the annual count of days where daily minimum
temperature drops below 0 degrees Celsius.
}
