% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.fg}
\alias{climdex.fg}
\title{Mean wind}
\usage{
climdex.fg(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean wind speed in [m/s])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the time series of mean wind speed.
}
\description{
This function computes the climdex index FG: the mean wind speed measured within a period.
}
\references{
\url{http://www.ecad.eu/documents/atbd.pdf}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
