% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.fg6bft}
\alias{climdex.fg6bft}
\title{Windy days}
\usage{
climdex.fg6bft(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean wind speed in [m/s])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the number of calm days.
}
\description{
This function computes the climdex index FG6bft: The number of days with mean wind greater than or equal 10.8 m/s (~6 Bft).
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
