% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.fgcalm}
\alias{climdex.fgcalm}
\title{Calm days}
\usage{
climdex.fgcalm(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean wind speed in [m/s])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the number of calm days.
}
\description{
This function computes the climdex index FGcalm days: The number of days with mean wind lower than or equal 2 m/s.
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
