% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.fxstorm}
\alias{climdex.fxstorm}
\title{Storm days}
\usage{
climdex.fxstorm(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily maximum wind gust in [m/s])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the number of storm days.
}
\description{
This function computes the climdex index FXstorm: The number of days with wind gusts greater than or equal 20.8 m/s (75 km/h).
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
