% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.fxx}
\alias{climdex.fxx}
\title{Maximum wind gust}
\usage{
climdex.fxx(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily maximum wind gust in [m/s])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the maximum of the maximum daily wind gust.
}
\description{
This function computes the climdex index FXx: The maximum of the maximum daily wind gust.
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
