% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{climdex.hd17}
\alias{climdex.hd17}
\title{Heating Degree Days}
\usage{
climdex.hd17(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{Object of type climdexInput. Here the daily maximum temperature.}

\item{freq}{Time frequency to aggregate to. Allowed only monthly, annual, halfyear, seasonal.}

\item{spells.can.span.years}{Default FALSE}
}
\value{
A vector containing a timeseries of the number of consecutive summer days in a given period (freq).
}
\description{
This function takes a climdexInput object as input and computes the climdex
index hd17: the annual (or at different periods) sum of heating degree days (17-tavg)
}
\author{
Christiana Photiadou (KNMI)
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}

