% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.id}
\alias{climdex.id}
\title{Icing Days}
\usage{
climdex.id(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the number of icing days for each year.
}
\description{
This function computes the climdex index ID.
}
\details{
This function takes a climdexInput object as input and computes the ID (icing
days) climdex index: that is, the annual count of days where daily maximum
temperature is below 0 degrees Celsius.
}
