% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.nsd}
\alias{climdex.nsd}
\title{Number of new snow days}
\usage{
climdex.nsd(ci, threshold = 1, freq = c("annual", "monthly", "halfyear",
  "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily snowfall timeseries in [cm]).}

\item{threshold}{Snow depth threshhold [in cm]. (Default: 1)}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear" or "seasonal". Default: "annual".}
}
\value{
A vector containing the timeseries of the number of new snow days.
}
\description{
This function computes the climdex index NSD: The number of days with a (new) snowfall > Y cm.
}
\author{
Rebekka Posselt (MeteoSwiss)
}
