% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.nsx}
\alias{climdex.nsx}
\title{Maximum new snow}
\usage{
climdex.nsx(ci, freq = c("annual", "monthly", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily snowfall timeseries in [cm]).}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear" or "seasonal". Default: "annual".}
}
\value{
A vector containing the timeseries of the maximum snowfall.
}
\description{
This function computes the climdex index NSX: The maximum snowfall measured within a year.
}
\author{
Rebekka Posselt (MeteoSwiss)
}
