% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.prcptot}
\alias{climdex.prcptot}
\title{Total Daily Precipitation}
\usage{
climdex.prcptot(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation in wet days.
}
\description{
This function computes the climdex index PRCPTOT.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index PRCPTOT: the annual sum of precipitation in wet days
(days where precipitation is at least 1mm).
}
