% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.r75p}
\alias{climdex.r75p}
\title{Total Daily Precipitation Exceeding 75\%ile Threshold}
\usage{
climdex.r75p(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation exceeding
the threshold.
}
\description{
This function computes the climdex index R75p.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R75p: the annual sum of precipitation in days where daily precipitation exceeds the
75th percentile of daily precipitation in the base period.
}
