% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.r75pfrac}
\alias{climdex.r75pfrac}
\title{Precipitation fraction due to moderate wet days (exceeding 75\%ile Threshold)}
\usage{
climdex.r75pfrac(ci, freq = c("monthly", "annual", "halfyear",
  "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation exceeding
the threshold.
}
\description{
This function computes the climdex index R75pFRAC.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R75pFRAC: the fraction of the sum of precipitation in days where daily precipitation exceeds the
75th percentile of daily precipitation in the base period.
}
