% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.r95p}
\alias{climdex.r95p}
\title{Total Daily Precipitation Exceeding 95\%ile Threshold}
\usage{
climdex.r95p(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation exceeding
the threshold.
}
\description{
This function computes the climdex index R95p.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R95p: the annual sum of precipitation in days where daily precipitation exceeds the
95th percentile of daily precipitation in the base period.
}
