% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.r95ptot}
\alias{climdex.r95ptot}
\title{Precipitation fraction due to very wet days (exceeding 95\%ile Threshold)}
\usage{
climdex.r95ptot(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation exceeding
the threshold.
}
\description{
This function computes the climdex index R95pTOT.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R95pTOT: the fraction of the sum of precipitation in days where daily precipitation exceeds the
95th percentile of daily precipitation in the base period.
}
