% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.r99pfrac}
\alias{climdex.r99pfrac}
\title{Precipitation fraction due to extremly wet days (exceeding 99\%ile Threshold)}
\usage{
climdex.r99pfrac(ci, freq = c("monthly", "annual", "halfyear",
  "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing an annual timeseries of precipitation exceeding
the threshold.
}
\description{
This function computes the climdex index R99pFRAC.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index R99pFRAC: the fraction of the sum of precipitation in days where daily precipitation exceeds the
99th percentile of daily precipitation in the base period.
}
