% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.rnnmm}
\alias{climdex.rnnmm}
\title{Precipitation Exceeding A Specified Amount Per Day}
\usage{
climdex.rnnmm(ci, threshold = 1, freq = c("monthly", "annual", "halfyear",
  "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{threshold}{The threshold to be used for Rnnmm.}
}
\value{
A vector containing the value of the index for each year.
}
\description{
This function computes the climdex index Rnnmm.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index Rnnmm: the annual count of days where daily precipitation is more than \code{nn} mm per day.
}
