% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.rx1day}
\alias{climdex.rx1day}
\title{Monthly Maximum 1-day Precipitation}
\usage{
climdex.rx1day(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}
}
\value{
A vector containing the value of the index for each month of each
year.
}
\description{
This function computes the climdex index Rx1day.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index Rx1day: monthly or annual maximum 1-day precipitation.
}
\note{
The default behaviour of climdex.rx5day differs somewhat from
fclimdex, as fclimdex and climdex.pcic differ on the definition of Rx5day.
The running sum series computed by fclimdex is off by 2 days, and the first
day a running sum can be computed for is left out entirely. The behaviour of
fclimdex can be replicated by setting center.mean.on.last.day to TRUE.
}
