% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.sdd}
\alias{climdex.sdd}
\title{Number of snow days}
\usage{
climdex.sdd(ci, threshold = 1, freq = c("annual", "monthly", "halfyear",
  "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily snow depth timeseries in [cm]).}

\item{threshold}{Snow depth threshhold [in cm]. (Default: 1)}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear" or "seasonal". Default: "annual".}
}
\value{
A vector containing an annual timeseries of the number of snow days.
}
\description{
This function computes the climdex index SDD: The number of days with a snow depth > Y cm.
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}
\author{
Rebekka Posselt (MeteoSwiss)
}
