% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.sdii}
\alias{climdex.sdii}
\title{Simple Precpitation Intensity Index}
\usage{
climdex.sdii(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the value of the index for each year.
}
\description{
This function computes the climdex index SDII.
}
\details{
\code{climdex.sdii} computes the climdex index SDII, or Simple Precipitation
Intensity Index. This is defined as the sum of precipitation in wet days
(days with preciptitation over 1mm) during the year divided by the number of
wet days in the year.
}
\note{
fclimdex rounds to 1 decimal place, whereas climdex.sdii does not.
This results in some small differences.
}
