% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.sdx}
\alias{climdex.sdx}
\title{Maximum snow depth}
\usage{
climdex.sdx(ci, freq = c("annual", "monthly", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily snow depth timeseries in [cm]).}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear" or "seasonal". Default: "annual".}
}
\value{
A vector containing an annual timeseries of the maximum snow depth.
}
\description{
This function computes the climdex index SDx: The maximum snow depth measured within a year.
}
\author{
Rebekka Posselt (MeteoSwiss)
}
