% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.spi3}
\alias{climdex.spi3}
\title{Standardized Precipitation Index 3-mon}
\usage{
climdex.spi3(ci, freq = c("monthly"), scale = 3, distribution = "Gamma",
  fit = "ub-pwm", kernal = list(type = "rectangular", shift = 1),
  ref.start = NULL, ref.end = NULL)
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily precipitation [mm])}

\item{freq}{Time frequency to aggregate to. Allowed only monthly}

\item{scale}{an integer, representing the time scale at which the SPI will be computed. Default is 3}

\item{distribution}{name of the distribution function to be used for computing the SPI (default ’Gamma’)}

\item{fit}{name of the method used for computing the distribution function parameters (default 'ub-pwm')}

\item{kernal}{optional, a list defining the type of kernel used for computing the SPI atscales higher than one. Defaults to unshifted rectangular kernel.}

\item{ref.start}{optional, starting point of the reference period used for computing the index. Defaults to NULL, indicating that the first value in data will be used as starting point.}

\item{ref.end}{optional, ending  point  of  the  reference  period  used  for  computing  the index. Defaults to NULL, indicating that the last value in data will be used as ending point.}

\item{...}{For more details please refer to the SPEI}
}
\value{
A vector containing a monthly SPI at the selected scale
}
\description{
Adapted by ClimPACT2 and SPEI (arguments taken from here) libraries.
Given a timeseries of daily precipitation amounts [mm], the function calculates the Standardized Precipitation Index (SPI) as it is created and defined by the SPEI library.
This function computes the climdex index spi at multiple scales.
}
\references{
\url{http://www.ecad.eu/documents/atbd.pdf}

\url{https://cran.r-project.org/web/packages/SPEI/SPEI.pdf}
}
\author{
Christiana Photiadou (KNMI)
}
