% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.ss}
\alias{climdex.ss}
\title{Sunshine duration}
\usage{
climdex.ss(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily sunshine duration in [h])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the time series of sunshine duration.
}
\description{
This function computes the climdex index SS: The sum of the sunshine duration hours within a period.
}
\author{
Rebekka Posselt (MeteoSwiss)
}
