% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.su}
\alias{climdex.su}
\title{Summer Days}
\usage{
climdex.su(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the number of summer days for each year.
}
\description{
This function computes the climdex index SU.
}
\details{
This function takes a climdexInput object as input and computes the SU (summer
days) climdex index: that is, the annual count of days where daily maximum
temperature exceeds 25 degrees Celsius.
}
