% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.sun_cloudy}
\alias{climdex.sun_cloudy}
\title{Mostly cloudy days}
\usage{
climdex.sun_cloudy(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily relative sunshine duration in [\%])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the number of mostly cloudy days.
}
\description{
This function computes the climdex index SUN_CLOUDY.

This function takes a climdexInput object as input and computes the 
SUN_CLOUDY index: The number of mostly cloudy days (relative sunshine duration < 20 \%) within a period.
}
