% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.sun_relmean}
\alias{climdex.sun_relmean}
\title{Relative sunshine duration}
\usage{
climdex.sun_relmean(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily relative sunshine duration in [\%])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the time series of the relative sunshine duration.
}
\description{
This function computes the climdex index SUN_RELMEAN.

This function takes a climdexInput object as input and computes the 
SUN_RELMEAN index: The mean of the relative sunshine duration within a period.
}
