% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.sun_sunny}
\alias{climdex.sun_sunny}
\title{Mostly sunny days}
\usage{
climdex.sun_sunny(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily relative sunshine duration in [\%])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the number of mostly sunny days.
}
\description{
This function computes the climdex index SUN_SUNNY.

This function takes a climdexInput object as input and computes the 
SUN_SUNNY index: The number of mostly sunny days (relative sunshine duration > 80 \%) within a period.
}
