% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.tnn}
\alias{climdex.tnn}
\title{Monthly Minimum of Daily Minimum Temperature}
\usage{
climdex.tnn(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}
}
\value{
A vector containing the value of the index for each month.
}
\description{
This function computes the climdex index TNn.
}
\details{
This function takes a climdexInput object as input and computes
the monthly or annual minimum of daily minimum temperature.
}
