% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.tr}
\alias{climdex.tr}
\title{Tropical Nights}
\usage{
climdex.tr(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}
}
\value{
A vector containing the number of frost days for each year.
}
\description{
This function computes the climdex index TR.
}
\details{
This function takes a climdexInput object as input and computes the TR
(tropical nights) climdex index: that is, the annual count of days where
daily minimum temperature stays above 20 degrees Celsius.
}
