% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.txn}
\alias{climdex.txn}
\title{Monthly Minimum of Daily Maximum Temperature}
\usage{
climdex.txn(ci, freq = c("monthly", "annual", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput.}

\item{freq}{Time frequency to aggregate to.}
}
\value{
A vector containing the value of the index for each month.
}
\description{
This function computes the climdex index TXn.
}
\details{
This function takes a climdexInput object as input and computes
the monthly or annual minimum of daily maximum temperature.
}
