% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{climdex.wd}
\alias{climdex.wd}
\title{Warm-dry days}
\usage{
climdex.wd(ci, freq = c("monthly", "annual", "halfyear", "seasonal"),
  precip.thresh = "q25", precip.op = "<", temp.thresh = "q75",
  temp.op = ">")
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily precipitation [mm] and the averaged daily temperature [C])}
}
\value{
A vector containing an annual timeseries of precipitation in wet days.
}
\description{
This function computes the climdex index WD.
}
\details{
This function takes a climdexInput object as input and computes the climdex
index WD: the number of days where TG>75 & RR<25 (for wet days: days where precipitation is at least 1mm).
}
\references{
\url{http://etccdi.pacificclimate.org/list_27_indices.shtml}
}
\seealso{
\code{\link{climdexInput.raw}}, \code{\link{climdexInput.csv}}.
}
\keyword{climate}
\keyword{ts}
