% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{cloudy_days}
\alias{cloudy_days}
\title{119. cloudy days}
\usage{
cloudy_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{cloud base below 100 meter}

\item{data_names}{names of each period of time
Number of days with cloud base below 100 meter.}

\item{time.scale}{month, season or year}
}
\value{
cloudy days
}
\description{
119. cloudy days
}
\examples{
calculate119(data = cloud_cover_less_100.value)
}
