% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{CMD}
\alias{CMD}
\title{96. Climatic Moisture Deficit (CMD)
ETo - Effective Precipitation}
\usage{
CMD(eto, pr, taverage, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{eto}{et0}

\item{pr}{precipitation}

\item{taverage}{taverage}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
CMD
}
\description{
96. Climatic Moisture Deficit (CMD)
ETo - Effective Precipitation
}
\examples{
calculate96(eto = eto.value, pr = pr.value)
}
