% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{col_ddays}
\alias{col_ddays}
\title{8. Cold days
Percentages of days with maximum temperatures lower than the 10th percentile.}
\usage{
col_ddays(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Cold days
}
\description{
8. Cold days
Percentages of days with maximum temperatures lower than the 10th percentile.
}
\examples{
calculate8(data=tmax.value)
}
