% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{cold_nights}
\alias{cold_nights}
\title{9. Cold nights
Percentages of days with minimum temperatures lower than the 10th percentile.}
\usage{
cold_nights(data, date = NA, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{date}{date}

\item{time.scale}{month, season or year}
}
\value{
Cold nights
}
\description{
9. Cold nights
Percentages of days with minimum temperatures lower than the 10th percentile.
}
\examples{
calculate9(data=tmin.value)
}
