% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{cold_spell_duration_index}
\alias{cold_spell_duration_index}
\title{10. Cold spell duration index
Annual count of days with at least 6 consecutive days when TN < 10th percentile #Tn = tmin}
\usage{
cold_spell_duration_index(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Cold spell duration index
}
\description{
10. Cold spell duration index
Annual count of days with at least 6 consecutive days when TN < 10th percentile #Tn = tmin
}
\examples{
calculate10(data=tmin.value)
}
