% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{consecutive_summer_days}
\alias{consecutive_summer_days}
\title{20. Maximum number of consecutive summer days (TX > 25º)
Monthly and annual}
\usage{
consecutive_summer_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
consecutive summer days
}
\description{
20. Maximum number of consecutive summer days (TX > 25º)
Monthly and annual
}
\examples{
calculate20(data=tmax.value)
}
