% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{csd}
\alias{csd}
\title{Maximum consecutive summer days}
\usage{
csd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
consecutive summer days
}
\description{
Maximum number of consecutive summer days (TX > 25 Celsius)
}
\examples{
data(data_all)
csd(data=data_all$tx)
}
