% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{d32}
\alias{d32}
\title{Days TX32}
\usage{
d32(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{daily maximum temperature, Celsius}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
days
}
\description{
Number of days whith TX >= 32 Celsius on the interval June-August.
}
\examples{
data(data_all)
d32(data = data_all$tx)
}
