% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{d50mm}
\alias{d50mm}
\title{Heavy precipitation days}
\usage{
d50mm(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{daily precipitation, mm}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
days
}
\description{
Number of days with precipitation above 50mm
}
\examples{
data(data_all)
d50mm(data = data_all$rr)
}
