% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/data.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{radiation.value}
\alias{toa.value}
\alias{MinMAP}
\title{radiation}
\format{An object of class \code{integer} of length 123.}
\usage{
data(radiation)

data(toa)

MinMAP
}
\description{
Data used in the examples

See wichita

See wichita
}
\details{
See description.

See description.

See description.
}
\keyword{datasets}
