% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{dd}
\alias{dd}
\title{Dry days}
\usage{
dd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
dry days
}
\description{
Number of days with less than 1 mm
}
\examples{
data(data_all)
dd(data = data_all$rr)
}
