% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{dfx21}
\alias{dfx21}
\title{Days wind gusts above 21 m/s}
\usage{
dfx21(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Gustmax
}
\description{
Number of days with wind gusts above 21 m/s
}
\examples{
data(data_all)
dfx21(data = data_all$windGUST)
}
