% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{diurnal_temperature_range}
\alias{diurnal_temperature_range}
\title{12. Diurnal temperature range
Mean difference between TX and TN.}
\usage{
diurnal_temperature_range(max, min, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{max}{maximum temperature}

\item{min}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Diurnal temperature range
}
\description{
12. Diurnal temperature range
Mean difference between TX and TN.
}
\examples{
calculate12(max=tmax.value, min=tmin.value)
}
