% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{dr1mm}
\alias{dr1mm}
\title{Wet days 1mm}
\usage{
dr1mm(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
RR1
}
\description{
Total number of wet days >= 1 mm
}
\examples{
data(data_all)
dr1mm(data = data_all$rr)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
