% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{dr3mm}
\alias{dr3mm}
\title{Wet days 3mm}
\usage{
dr3mm(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
RR3
}
\description{
Total number of Wet days >= 3mm
}
\examples{
data(data_all)
dr3mm(data = data_all$rr)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
