% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{dry_days}
\alias{dry_days}
\title{53. Dry days
The annual number of days with less than 1 mm/day}
\usage{
dry_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
dry days
}
\description{
53. Dry days
The annual number of days with less than 1 mm/day
}
\examples{
calculate53(data = pr.value)
}
