% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{eca.input}
\alias{eca.input}
\title{ECA &D station data files}
\usage{
eca.input(filename, var.name, date.name)
}
\arguments{
\item{filename}{File name and path of station data file.}

\item{var.name}{A varialbe name from the ECA\&D variavle list: prec (RR), tavg (TG), tmax (TX), tmin(TN), sun (SS),
wind_gust (FX), wind (FG).}

\item{data.name}{Always DATE}
}
\value{
A data frame containing two columns: DATE with dates in PCICt format and "var.name" the varialbe name.
}
\description{
This function reads and prepares the station data files from the European Climate Assessment and Dataset (ECA&D) for use in climdex
}
\examples{
library(PCICt)

## Create a climdexInput object from some data already loaded in and
## ready to go.

## Parse the dates into PCICt.
tmax.dates <- as.PCICt(do.call(paste, ec.1018935.tmax[,c("year",
"jday")]), format="\%Y \%j", cal="gregorian")
tmin.dates <- as.PCICt(do.call(paste, ec.1018935.tmin[,c("year",
"jday")]), format="\%Y \%j", cal="gregorian")
prec.dates <- as.PCICt(do.call(paste, ec.1018935.prec[,c("year",
"jday")]), format="\%Y \%j", cal="gregorian")

## Load the data in.
ci <- climdexInput.raw(ec.1018935.tmax$MAX_TEMP,
ec.1018935.tmin$MIN_TEMP, ec.1018935.prec$ONE_DAY_PRECIPITATION,
tmax.dates, tmin.dates, prec.dates, base.range=c(1971, 2000))

## Create 
}
\author{
Christiana Photiadou (KNMI)
}
\references{
\url{http://www.ecad.eu/}
}

