% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{effective_precipitation}
\alias{effective_precipitation}
\title{55. Effective precipitation
precipitation minus evapotranspiration}
\usage{
effective_precipitation(pr, eto, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{pr}{precipitation}

\item{eto}{et0}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
effective precipitation
}
\description{
55. Effective precipitation
precipitation minus evapotranspiration
}
\examples{
calculate55(pr = pr.value, eto = eto.value)
}
