% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{emberger_aridity_index}
\alias{emberger_aridity_index}
\title{98. Emberger Aridity Index
 100*P / (M^2 -  m^2)}
\usage{
emberger_aridity_index(taverage, pr, data_names = NULL)
}
\arguments{
\item{taverage}{medium temperature}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}
}
\value{
Emberger Aridity Index
}
\description{
98. Emberger Aridity Index
 100*P / (M^2 -  m^2)
}
\examples{
calculate98(taverage = taverage.value, pr = pr.value)
}
