% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{Eto}
\alias{Eto}
\title{94. Annual, seasonal, and monthly Reference Evapotranspiration (Eto)
If data available using Fao-56 Penman-Monteith, if not using the Hargreaves & Samani method.}
\usage{
Eto(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{Eto}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Eto
}
\description{
94. Annual, seasonal, and monthly Reference Evapotranspiration (Eto)
If data available using Fao-56 Penman-Monteith, if not using the Hargreaves & Samani method.
}
\examples{
calculate94(data = eto.value)
}
