% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{extreme_temperature_range}
\alias{extreme_temperature_range}
\title{18. Intra-annual extreme temperature range
Difference between the highest TX and the lowest TN in the year.}
\usage{
extreme_temperature_range(max, min, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{max}{maximum temperature}

\item{min}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
extreme temperature range
}
\description{
18. Intra-annual extreme temperature range
Difference between the highest TX and the lowest TN in the year.
}
\examples{
calculate18(max=tmax.value, min=tmin.value)
}
