% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FFDI}
\alias{FFDI}
\title{128. McArthur Forest Fire Danger Index (FFDI)
combination of temperature, relative humidity, surface wind speed and KBDI
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R#Keech-Byran_Drought_Index}
\usage{
FFDI(taverage, pr, rh, w, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{pr}{precipitation}

\item{rh}{relative humidity}

\item{w}{average wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FFDI
}
\description{
128. McArthur Forest Fire Danger Index (FFDI)
combination of temperature, relative humidity, surface wind speed and KBDI
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R#Keech-Byran_Drought_Index
}
\examples{
calculate128(taverage = taverage.value, pr=pr.value, rh=rh.value, w=w.value)
}
