% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macArthurFFDI.R
\name{ffdiIndex}
\alias{ffdiIndex}
\title{Mac Arthur Forest Fire Danger Index}
\usage{
ffdiIndex(madf, t, h, w)
}
\arguments{
\item{madf}{Mac Arthur's Drought Index. This is the output of the \code{\link{kbdindex}} function, using the option \code{what = "madf"}.}

\item{t}{t}

\item{h}{Vector of relative humidity data (in percentage)}

\item{w}{Vector of wind velocity records (in km/h)}
}
\value{
A vector of (daily) FFDI data
}
\description{
Implementation of the Mac Arthur's Forest Fire Danger Index (FFDI)
}
\references{
McArthur, A.G. (1973) Forest Fire Danger Meter Mk.5. Commonwealth of Australian Forestry and Timber Bureau.
}
\seealso{
kbdindex
}
\author{
Joaquin Bedia-Jiménez
}
\keyword{internal}
