% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FFFI}
\alias{FFFI}
\title{130. Finnish Forest Fire Index (FFFI)
combination of temperature, relative humidity, wind speed, radiation and precipitation
https://link.springer.com/chapter/10.1007%2F978-3-642-55903-7_88
P-eto+-f1}
\usage{
FFFI(data, toa, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{net radiation}

\item{toa}{solar radiation at TOA}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FFFI
}
\description{
130. Finnish Forest Fire Index (FFFI)
combination of temperature, relative humidity, wind speed, radiation and precipitation
https://link.springer.com/chapter/10.1007%2F978-3-642-55903-7_88
P-eto+-f1
}
\examples{
calculate130(data = radiation.value, toa=toa.value)
}
