% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FG}
\alias{FG}
\title{91. FG:
Mean of daily mean wind strength (m/s), ECA&D standard}
\usage{
FG(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FG
}
\description{
91. FG:
Mean of daily mean wind strength (m/s), ECA&D standard
}
\examples{
calculate91(data = w.value)
}
