% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FG6Bft}
\alias{FG6Bft}
\title{93. FG6Bft: 
Days with daily averaged wind >= 6 Bft (10.8 m/s) (days), ECA&D standard}
\usage{
FG6Bft(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FG6Bft
}
\description{
93. FG6Bft: 
Days with daily averaged wind >= 6 Bft (10.8 m/s) (days), ECA&D standard
}
\examples{
calculate93(data = w.value)
}
