% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FGcalm}
\alias{FGcalm}
\title{92. FGcalm: 
Calm days (FG <= 2 m/s) (days), ECA&D standard}
\usage{
FGcalm(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FGcalm
}
\description{
92. FGcalm: 
Calm days (FG <= 2 m/s) (days), ECA&D standard
}
\examples{
calculate92(data = w.value)
}
