% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{first_permanent_snowcover}
\alias{first_permanent_snowcover}
\title{112. The arrival date of first permanent snowcover
The first date of a snow spell (consecutive snow cover days) in the year.}
\usage{
first_permanent_snowcover(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
first permanent snowcover
}
\description{
112. The arrival date of first permanent snowcover
The first date of a snow spell (consecutive snow cover days) in the year.
}
\examples{
calculate112(data = snow_depth.value)
}
