% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{first_snowcover}
\alias{first_snowcover}
\title{111. The arrival date of first snowcover}
\usage{
first_snowcover(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
first snowcover
}
\description{
111. The arrival date of first snowcover
}
\examples{
calculate111(data = snow_depth.value)
}
