% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{fod}
\alias{fod}
\title{Foggy days}
\usage{
fod(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{cloud base below 100 meter}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
cloudy days
}
\description{
Number of days with fog.
}
\examples{
data(data_all)
fod(data = data_all$cloud100)
}
\references{
Rastogi, B., A.P. Williams, D.T. Fischer, S.F. Iacobellis, K. McEachern, L. Carvalho, C. Jones, S.A. Baguskas, and C.J. Still, 2016: Spatial and Temporal Patterns of Cloud Cover and Fog Inundation in Coastal California: Ecological Implications. Earth Interact., 20, 1–19, \url{https://doi.org/10.1175/EI-D-15-0033.1}
}
