% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_ind_MCH.R
\name{fraction.within.cons.op.threshold}
\alias{fraction.within.cons.op.threshold}
\title{Fraction of days which are within spells of at least n+1 days}
\usage{
fraction.within.cons.op.threshold(cloud.bin, date.factor, n)
}
\arguments{
\item{cloud.bin}{Time series of binary data (counted are 1s)}

\item{date.factor}{Factor to aggregate by.}

\item{n}{Maximum number of spell lenghts that should be skipped.}
}
\value{
for each date.factor the fraction of days which are whithin
at least n+1 consecutive days
}
\description{
This function calculates the fraction of days which are within spells with lengths of at least n+1 days.
It uses two functions, select.blocks.gt.length and number.days.op.threshold, which
were found in https://github.com/ECA-D/climind/blob/master/R/climdex.r
}
