% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{freezing_rain}
\alias{freezing_rain}
\title{54. Freezing rain
The annual number of days with freezing rain}
\usage{
freezing_rain(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{freezing rain}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
freezing rain days
}
\description{
54. Freezing rain
The annual number of days with freezing rain
}
\examples{
calculate54(data = freezing.rain.value)
}
