% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{freq_snow_days}
\alias{freq_snow_days}
\title{108. Freq. of snow days
annual number of snow days}
\usage{
freq_snow_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snowfall}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
freq. of snow days
}
\description{
108. Freq. of snow days
annual number of snow days
}
\examples{
calculate108(data = snow.value)
}
