% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{frost_days}
\alias{frost_days}
\title{14. Frost days
Number of days with minimum temperature <0ºC per year.}
\usage{
frost_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
frost days
}
\description{
14. Frost days
Number of days with minimum temperature <0ºC per year.
}
\examples{
calculate14(data=tmin.value)
}
