% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{fsc}
\alias{fsc}
\title{Date of first snow cover}
\usage{
fsc(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{snow depth, mm}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{further arguments passed to or from other methods}
}
\value{
date
}
\description{
First day when there is measurable snow cover (day of the hydrological year)
}
\examples{
data(data_all)
fsc(data = data_all$snowdepth)
}
