% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{fsd}
\alias{fsd}
\title{Number of snow days}
\usage{
fsd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snowfall}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
freq. of snow days
}
\description{
Number of snow days
number of snow days
No. snow days
}
\examples{
data(data_all)
fsd(data = data_all$snowfall)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
