% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FWI}
\alias{FWI}
\title{126. Canadian Fire Weather Index (FWI)
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R
https://github.com/SantanderMetGroup/fireDanger
combination of daily values of temperature, relative humidity, surface wind and precipitation}
\usage{
FWI(taverage, rh, w, pr, dew_point, lat, data_names = NULL,
  time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{rh}{relative humidity}

\item{w}{average wind}

\item{pr}{precipitation}

\item{dew_point}{dew_point}

\item{lat}{latitude}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FWI
}
\description{
126. Canadian Fire Weather Index (FWI)
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R
https://github.com/SantanderMetGroup/fireDanger
combination of daily values of temperature, relative humidity, surface wind and precipitation
}
\examples{
calculate126(taverage=taverage.value, rh = rh.value, w = w.value, 
pr = pr.value, lat = lat, dew_point=dew_point.value)
}
