% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/fwi1D.R
\name{fwi1D}
\alias{fwi1D}
\title{Fire Weather Index}
\usage{
fwi1D(months, Tm, H, r, W, lat = 46, return.all = FALSE, init.pars = c(85,
  6, 15))
}
\arguments{
\item{months}{months}

\item{Tm}{Vector of temperature records (deg. Celsius)}

\item{H}{Vector of relative humidity records (\%)}

\item{r}{Vector of last 24-h accumulated precipitation (mm)}

\item{W}{Vector of wind velocity records (Km/h)}

\item{lat}{Optional. Latitude of the records (in decimal degrees). Default to 46,
applying the default parameters of the original FWI System, developed in Canada. See details.}

\item{return.all}{Logical. Should all components of the FWI system be returned?. 
Default to FALSE, indicating that only FWI is returned.}

\item{init.pars}{A numeric vector of length 3 with the initialization values for the
FFMC, DMC and DC components, in this order. Default values as proposed by van Wagner (1987).}
}
\value{
A vector of the same length as the input vectors (minus possible missing observations),
containing the requested components of the FWI system (either all or just FWI). See details.
}
\description{
Implementation of the Canadian Fire Weather Index System
}
\section{Daylength adjustment factors}{
 
By default, the function applies the original FWI daylength adjustment factors for DC and DMC (van Wagner 1987),
 although it is possible to adjust them by as a function of latitude through the argument \code{lat}.
The reference values used for each latitudinal range are those indicated in p.71 and Tables A3.1 and A3.2 (p69) in
Lawson and Armitage (2008).
}

\references{
Lawson, B.D. & Armitage, O.B., 2008. Weather guide for the Canadian Forest Fire Danger Rating System. Northern Forestry Centre, Edmonton (Canada).

van Wagner, C.E., 1987. Development and structure of the Canadian Forest Fire Weather Index (Forestry Tech. Rep. No. 35). Canadian Forestry Service, Ottawa, Canada.

van Wagner, C.E., Pickett, T.L., 1985. Equations and FORTRAN program for the Canadian forest fire weather index system (Forestry Tech. Rep. No. 33). Canadian Forestry Service, Ottawa, Canada.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}, partially based on the original FORTRAN code by van Wagner and Pickett (1985)
}
