% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{FXx}
\alias{FXx}
\title{90. FXx:
Maximun value of daily maximum wind gust (m/s), ECA&D standard}
\usage{
FXx(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum wind gust}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
FXx
}
\description{
90. FXx:
Maximun value of daily maximum wind gust (m/s), ECA&D standard
}
\examples{
calculate90(data = w_max.value)
}
