% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{GD4}
\alias{GD4}
\title{35. Growing degree days (GD4)
Sum of degree days over 4ºC (annual and monthly) (sum Cº-4)}
\usage{
GD4(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
GD4
}
\description{
35. Growing degree days (GD4)
Sum of degree days over 4ºC (annual and monthly) (sum Cº-4)
}
\examples{
calculate35(data=taverage.value)
}
