% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{growing_precipitation}
\alias{growing_precipitation}
\title{64. Growing season precipitation
Growing season (april to october) mean precipitation}
\usage{
growing_precipitation(data, data_names = NULL)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}
}
\value{
growing precipitation
}
\description{
64. Growing season precipitation
Growing season (april to october) mean precipitation
}
\examples{
calculate64(data = pr.value)
}
