% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{growing_season_length}
\alias{growing_season_length}
\title{15. Growing season length
Annual count of days between the first span of at least 6 days with Tmean >5ºC and first span after 1 July of 6 days with Tmean <5 ºC.}
\usage{
growing_season_length(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{mean temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
growing season length
}
\description{
15. Growing season length
Annual count of days between the first span of at least 6 days with Tmean >5ºC and first span after 1 July of 6 days with Tmean <5 ºC.
}
\examples{
calculate15(data=taverage.value)
}
