% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{growing_temp_1}
\alias{growing_temp_1}
\title{33. Growing season temperature 1}
\usage{
growing_temp_1(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time
Growing season (april to october) mean temperature}
}
\value{
Growing season temperature 1
}
\description{
33. Growing season temperature 1
}
\examples{
calculate33(data=taverage.value)
}
