% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{growing_temp_2}
\alias{growing_temp_2}
\title{34. Growing season temperature 2
Growing season (may to september) mean temperature}
\usage{
growing_temp_2(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
Growing season temperature 2
}
\description{
34. Growing season temperature 2
Growing season (may to september) mean temperature
}
\examples{
calculate34(data=taverage.value)
}
