% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{gtx}
\alias{gtx}
\title{Mean TX}
\usage{
gtx(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Average temperature
}
\description{
Mean of daily maximum air temperature
}
\examples{
data(data_all)
gtx(data=data_all$tg)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
