% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{Gustmax}
\alias{Gustmax}
\title{89. Gustmax
number of days with wind gusts above 21 m/s}
\usage{
Gustmax(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Gustmax
}
\description{
89. Gustmax
number of days with wind gusts above 21 m/s
}
\examples{
calculate89(data = w.value)
}
