% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{HCI}
\alias{HCI}
\title{131. HCI:Urban
Holliday Climate Index for Urban destinations (Scott et all, 2016) (Tmax,wind,cloudiness,RH, precipitation) Scott, D., Rutty, M., Amelung, B. and Tang, M. (2016): An inter-comparison of the Holiday Climate Index (HCI) and the Tourism Climate Index (TCI), Atmosphere, 7, 80, doi:10.3390/atmos7060080}
\usage{
HCI(pr, w, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{pr}{precipitation}

\item{w}{average wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
HCI
}
\description{
131. HCI:Urban
Holliday Climate Index for Urban destinations (Scott et all, 2016) (Tmax,wind,cloudiness,RH, precipitation) Scott, D., Rutty, M., Amelung, B. and Tang, M. (2016): An inter-comparison of the Holiday Climate Index (HCI) and the Tourism Climate Index (TCI), Atmosphere, 7, 80, doi:10.3390/atmos7060080
}
\examples{
calculate131(pr = pr.value, w=w.value)
}
