% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{HD17}
\alias{HD17}
\title{23. Heating degree days (HD17)
Monthly and annual (sum(Cº-17))}
\usage{
HD17(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
HD17
}
\description{
23. Heating degree days (HD17)
Monthly and annual (sum(Cº-17))
}
\examples{
calculate23(data=tmin.value)
}
