% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{heat_index}
\alias{heat_index}
\title{86. Heat Index: temperature + humidity 
 http://www.wpc.ncep.noaa.gov/html/heatindex_equation.shtml ; available from the R package weathermetrics}
\usage{
heat_index(taverage, rh, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{rh}{relative humidity}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Heat Index
}
\description{
86. Heat Index: temperature + humidity 
 http://www.wpc.ncep.noaa.gov/html/heatindex_equation.shtml ; available from the R package weathermetrics
}
\examples{
calculate86(taverage = taverage.value, rh = rh.value)
}
