% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{heavy_precipitation_days}
\alias{heavy_precipitation_days}
\title{60. Heavy precipitation days
number of days with precipitation above 50mm}
\usage{
heavy_precipitation_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
heavy precipitation days
}
\description{
60. Heavy precipitation days
number of days with precipitation above 50mm
}
\examples{
calculate60(data = pr.value)
}
