% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{heavy_snowy_days}
\alias{heavy_snowy_days}
\title{110. heavy snowy days
annual number of days with snow depth more than 50 cm.}
\usage{
heavy_snowy_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
heavy snowy days
}
\description{
110. heavy snowy days
annual number of days with snow depth more than 50 cm.
}
\examples{
calculate110(data = snow_depth.value)
}
