% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{hsd}
\alias{hsd}
\title{Heavy snowy days}
\usage{
hsd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
heavy snowy days
}
\description{
Number of days with snow depth more than 50 cm.
}
\examples{
data(data_all)
hsd(data = data_all$snowdepththickness)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
