% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{hydrological_years}
\alias{hydrological_years}
\title{Hydrological years}
\usage{
hydrological_years(time)
}
\arguments{
\item{time}{chron}
}
\value{
seasonals by years
}
\description{
Hydrological years
}
\keyword{internal}
