% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{ice_days}
\alias{ice_days}
\title{16. Ice days
Number of days with maximum temperature <0ºC per year.}
\usage{
ice_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
ice days
}
\description{
16. Ice days
Number of days with maximum temperature <0ºC per year.
}
\examples{
calculate16(data=tmax.value)
}
