% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis-package.R
\docType{package}
\name{indecis}
\alias{indecis}
\alias{indecis-package}
\title{indecis}
\description{
indecis
}
\author{
Fergus Reig Gracia 
Maintainer: Fergus Reig Gracia
\url{https://github.com/indecis-eu/indecis}
}
\keyword{package}
