% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{iUNEP}
\alias{iUNEP}
\title{95. iUNEP Aridity Index
P/Eto}
\usage{
iUNEP(eto, pr, taverage, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{eto}{et0}

\item{pr}{precipitation}

\item{taverage}{taverage}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
iUNEP
}
\description{
95. iUNEP Aridity Index
P/Eto
}
\examples{
calculate95(eto = eto.value, pr = pr.value)
}
