% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{JCI}
\alias{JCI}
\title{99. Johansson Continentality Index
 JCI = ( 1.7 * the annual range of monthly mean air temperatures grados / sin(geographic latitude grados) ) - 20.4}
\usage{
JCI(data, data_names = NULL, value)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{value}{lat}
}
\value{
JCI
}
\description{
99. Johansson Continentality Index
 JCI = ( 1.7 * the annual range of monthly mean air temperatures grados / sin(geographic latitude grados) ) - 20.4
}
\examples{
calculate99(data = taverage.value, value = lat)
}
