% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{KBDI}
\alias{KBDI}
\title{127. Keetch-Byran Drought Index (KBDI)
combination of daily maximum in temperature and precipitation
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R#Keech-Byran_Drought_Index}
\usage{
KBDI(taverage, pr, rh, w, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{pr}{precipitation}

\item{rh}{relative humidity}

\item{w}{average wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
KBDI
}
\description{
127. Keetch-Byran Drought Index (KBDI)
combination of daily maximum in temperature and precipitation
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R#Keech-Byran_Drought_Index
}
\examples{
calculate127(taverage = taverage.value, rh=rh.value, w = w.value, pr=pr.value)
}
