% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{kerner_oceanity_index}
\alias{kerner_oceanity_index}
\title{100. Kerner Oceanity Index
 k = 100 * (temperatura media octubre en grados - temperatura media abril en grados) / the annual range of monthly mean air temperatures grados}
\usage{
kerner_oceanity_index(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
Kerner Oceanity Index
}
\description{
100. Kerner Oceanity Index
 k = 100 * (temperatura media octubre en grados - temperatura media abril en grados) / the annual range of monthly mean air temperatures grados
}
\examples{
calculate100(data = taverage.value)
}
