% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{last_permanent_snowcover}
\alias{last_permanent_snowcover}
\title{113. The departure date of last permanent snowcover
The last date of a snow spell (consecutive snow cover days) in the year.}
\usage{
last_permanent_snowcover(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
last permanent snowcover
}
\description{
113. The departure date of last permanent snowcover
The last date of a snow spell (consecutive snow cover days) in the year.
}
\examples{
calculate113(data = snow_depth.value)
}
