% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{longest_dry_period}
\alias{longest_dry_period}
\title{56. Longest dry period
maximum length of consecutive dry days (pp<1)}
\usage{
longest_dry_period(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
longest dry period
}
\description{
56. Longest dry period
maximum length of consecutive dry days (pp<1)
}
\examples{
calculate56(data = pr.value)
}
