% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{longest_wet_period}
\alias{longest_wet_period}
\title{57. Longest wet period}
\usage{
longest_wet_period(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time
maximum length of consecutive wet days (pp>=1)}

\item{time.scale}{month, season or year}
}
\value{
longest wet period
}
\description{
57. Longest wet period
}
\examples{
calculate57(data = pr.value)
}
