% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{lpsc}
\alias{lpsc}
\title{Date of last permanent snow cover}
\usage{
lpsc(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
last permanent snowcover
}
\description{
Last day of the longest period with consecutive snow cover day (day of the hydrological year).
}
\examples{
data(data_all)
lpsc(data = data_all$snowdepth)
}
