% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{martonne}
\alias{martonne}
\title{97. De Martonne Aridity Index i
 De Martonne = P / (T + 10); P is the annual total amount of precipitation (mm) and T is the mean annual air temperature (°C)}
\usage{
martonne(pr, taverage, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{pr}{precipitation}

\item{taverage}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Martonne Aridity Index
}
\description{
97. De Martonne Aridity Index i
 De Martonne = P / (T + 10); P is the annual total amount of precipitation (mm) and T is the mean annual air temperature (°C)
}
\examples{
calculate97(pr = pr.value, taverage = taverage.value)
}
