% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{maxf}
\alias{maxf}
\title{max with na.rm TRUE}
\usage{
maxf(..., na.rm = TRUE)
}
\arguments{
\item{...}{...}

\item{na.rm}{na.rm}
}
\value{
max
}
\description{
max with na.rm TRUE
}
\keyword{internal}
