% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{maximum_consecutive_frost}
\alias{maximum_consecutive_frost}
\title{17. Maximum number of consecutive frost days (CFD)
Annual and monthly}
\usage{
maximum_consecutive_frost(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
maximum consecutive frost
}
\description{
17. Maximum number of consecutive frost days (CFD)
Annual and monthly
}
\examples{
calculate17(data=tmin.value)
}
