% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{maximum_precipitation}
\alias{maximum_precipitation}
\title{45. Maximum precipitation of the month, season and year}
\usage{
maximum_precipitation(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
maximum precipitation
}
\description{
45. Maximum precipitation of the month, season and year
}
\examples{
calculate45(data = pr.value)
}
