% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{maximum_snow_depth}
\alias{maximum_snow_depth}
\title{116. Maximum snow depth of the year / season / month}
\usage{
maximum_snow_depth(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
maximum snow depth
}
\description{
116. Maximum snow depth of the year / season / month
}
\examples{
calculate116(data = snow_depth.value)
}
