% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{maximum_temp}
\alias{maximum_temp}
\title{3. Maximum temperature (for maximum, minimum and mean)}
\usage{
maximum_temp(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum, minimum or medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
maximum temperature
}
\description{
3. Maximum temperature (for maximum, minimum and mean)
}
\examples{
calculate3(data=tmax.value)
}
