% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{mean_CC}
\alias{mean_CC}
\title{120. mean CC
mean daily cloud cover}
\usage{
mean_CC(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{cloud cover}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
mean CC
}
\description{
120. mean CC
mean daily cloud cover
}
\examples{
calculate120(data = cloud_cover.value)
}
