% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{mean_tg}
\alias{mean_tg}
\title{1. Average mean temperature}
\usage{
mean_tg(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
average temperature
}
\description{
1. Average mean temperature
}
\examples{
mean_tg(data=tmean.value)
}
