% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{mean_tn}
\alias{mean_tn}
\title{1. Average minimum temperature}
\usage{
mean_tn(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
average temperature
}
\description{
1. Average minimum temperature
}
\examples{
mean_tn(data=tmin.value)
}
