% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{mean_tx}
\alias{mean_tx}
\title{1. Average maximum temperature}
\usage{
mean_tx(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
average temperature
}
\description{
1. Average maximum temperature
}
\examples{
mean_tx(data=tmax.value)
}
