% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/custom_functions.R
\name{meanf}
\alias{meanf}
\title{mean with na.rm TRUE}
\usage{
meanf(..., na.rm = TRUE)
}
\arguments{
\item{...}{...}

\item{na.rm}{na.rm}
}
\value{
mean
}
\description{
mean with na.rm TRUE
}
