% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{MFI}
\alias{MFI}
\title{63. iModified Fournier Index
a precipitation concentration index
https://es.scribd.com/document/76414093/modified-fournier-index}
\usage{
MFI(data, data_names = NULL)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}
}
\value{
iModified Fournier Index
}
\description{
63. iModified Fournier Index
a precipitation concentration index
https://es.scribd.com/document/76414093/modified-fournier-index
}
\examples{
calculate63(data = pr.value)
}
