% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{mi}
\alias{mi}
\title{Mould index}
\usage{
mi(taverage, rh, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{taverage}{daily mean temperature, Celsius}

\item{rh}{relative humidity, percentage}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
days
}
\description{
Number of days with  a relative humidity over 90% in combination with TG above 10 Celsius
}
\examples{
data(data_all)
mi(taverage = data_all$tg, rh = data_all$humidity)
}
