% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{mild_snowy_days}
\alias{mild_snowy_days}
\title{109. mild snowy days}
\usage{
mild_snowy_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time
annual number of days with snow depth more than 5 cm.}

\item{time.scale}{month, season or year}
}
\value{
mild snowy days
}
\description{
109. mild snowy days
}
\examples{
calculate109(data = snow_depth.value)
}
