% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{minimum_temp}
\alias{minimum_temp}
\title{4. Minimum temperature (for maximum, minimum and mean)}
\usage{
minimum_temp(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum, minimum or medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
minimum temperature
}
\description{
4. Minimum temperature (for maximum, minimum and mean)
}
\examples{
calculate4(data=tmax.value)
}
