% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{MNI}
\alias{MNI}
\title{129. Modified Nesterov Index (MNI)
cummulative function of temperature and dew point deficit
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R
https://github.com/jbedia/fireDanger/wiki/nesterovIndex}
\usage{
MNI(dew_point, taverage, rh, pr, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{dew_point}{dew point}

\item{taverage}{medium temperature}

\item{rh}{relative humidity}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
MNI
}
\description{
129. Modified Nesterov Index (MNI)
cummulative function of temperature and dew point deficit
http://www.atriplex.info/index.php/Fire_Danger_Index_Functions_in_R
https://github.com/jbedia/fireDanger/wiki/nesterovIndex
}
\examples{
calculate129(dew_point=dew_point.value, rh=rh.value, taverage=taverage.value, pr=pr.value)
}
