% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{MOI}
\alias{MOI}
\title{103. Marsz Oceanity Index
 MOI = ( 0.731 * geographic latitude grados + 1.767 ) / the annual range of monthly mean air temperatures grados}
\usage{
MOI(data, data_names = NULL, value)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{value}{lat}
}
\value{
n0to10
}
\description{
103. Marsz Oceanity Index
 MOI = ( 0.731 * geographic latitude grados + 1.767 ) / the annual range of monthly mean air temperatures grados
}
\examples{
calculate103(data = taverage.value, value = lat)
}
