% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{months_quarter}
\alias{months_quarter}
\title{Select quarter days}
\usage{
months_quarter(functionValues, selectFunction, selectValues,
  na.rm = FALSE)
}
\arguments{
\item{functionValues}{functionValues}

\item{selectFunction}{selectFunction}

\item{selectValues}{selectValues}

\item{na.rm}{na.rm}
}
\value{
quarter days
}
\description{
Select quarter days
}
\keyword{internal}
