% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{months_years}
\alias{months_years}
\title{Months by years}
\usage{
months_years(time)
}
\arguments{
\item{time}{chron}
}
\value{
months by years
}
\description{
Months by years
}
\examples{
months_years(time=chron(names(tmax.value)))
}
