% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{mould_index}
\alias{mould_index}
\title{85. Mould index
Number of days with  a relative humidity over 90\% in combination with temperatures above 10°C}
\usage{
mould_index(taverage, rh, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{rh}{relative humidity}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Mould index
}
\description{
85. Mould index
Number of days with  a relative humidity over 90\% in combination with temperatures above 10°C
}
\examples{
calculate85(taverage = taverage.value, rh = rh.value)
}
