% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ms}
\alias{ms}
\title{Maximum snow depth}
\usage{
ms(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snow depth, mm}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
snow depth, m
}
\description{
Maximum snow depth (m)
}
\examples{
data(data_all)
ms(data = data_all$snowdepth)
}
