% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{n0to10}
\alias{n0to10}
\title{105. Snow depth n0to10
The number of days with snow depth in the range 1-10 cm}
\usage{
n0to10(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
n0to10
}
\description{
105. Snow depth n0to10
The number of days with snow depth in the range 1-10 cm
}
\examples{
calculate105(data = snow_depth.value)
}
