% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{n10to20}
\alias{n10to20}
\title{106. Snow depth n10to20
The number of days with snow depth of 10-20 cm}
\usage{
n10to20(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
n10to20
}
\description{
106. Snow depth n10to20
The number of days with snow depth of 10-20 cm
}
\examples{
calculate106(data = snow_depth.value)
}
