% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ngsr}
\alias{ngsr}
\title{Non-growing season precipitation}
\usage{
ngsr(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{daily precipitation, mm}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{further arguments passed to or from other methods}
}
\value{
precipitation, mm
}
\description{
Total precipitation from October to April
}
\examples{
data(data_all)
ngsr(data = data_all$rr)
}
