% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis.R
\name{no_null}
\alias{no_null}
\title{Select no empty parameters}
\usage{
no_null(data)
}
\arguments{
\item{data}{data list}
}
\value{
no empty parameter
}
\description{
Select no empty parameters
}
\keyword{internal}
