% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{non_growing_precipitation}
\alias{non_growing_precipitation}
\title{65. Non-growing season precipitation
October to april total precipitation, can inform on the resource available for low potential evaporation conditions}
\usage{
non_growing_precipitation(data, data_names = NULL)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}
}
\value{
non growing precipitation
}
\description{
65. Non-growing season precipitation
October to april total precipitation, can inform on the resource available for low potential evaporation conditions
}
\examples{
calculate65(data = pr.value)
}
