% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ntg}
\alias{ntg}
\title{Minimum TG}
\usage{
ntg(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{daily mean temperature, Celsius}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Average temperature
}
\description{
Minimum value of daily mean air temperature
}
\examples{
data(data_all)
ntg(data=data_all$tg)
}
