% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ogs6}
\alias{ogs6}
\title{Onset of growing season 6 days}
\usage{
ogs6(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Onset of growing season 1
}
\description{
Date of the start of the first span with at least 6 days with TG >5 Celsius
}
\examples{
data(data_all)
ogs6(data=data_all$tg)
}
