% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{onset_growing_1}
\alias{onset_growing_1}
\title{31. Onset of growing season 1
The start of the first span with at least 6 days with Tmean >5ºC}
\usage{
onset_growing_1(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Onset of growing season 1
}
\description{
31. Onset of growing season 1
The start of the first span with at least 6 days with Tmean >5ºC
}
\examples{
calculate31(data=taverage.value)
}
