% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_fao_dia.R
\name{penman_rs}
\alias{penman_rs}
\title{Transforma datos de in en r o al revés}
\usage{
penman_rs(J, lat = NA, tsun = NA, z = NA, ret = RADIATION)
}
\arguments{
\item{J}{Días de inicio de cada semana del año, partiendo desde 0 ¿?}

\item{lat}{latitud en grados en spTransform(coordenadas,CRS(crslonlat))}

\item{tsun}{Insolación en horas de sol o radiación en ¿MJ/m2?}

\item{z}{mde, modelo de elevación digital del terreno}

\item{ret}{Que hacer, calcular in desde r o al contrario}
}
\value{
insolación en horas de sol o radiación en ¿MJ/m2?
}
\description{
Transforma datos de in en r o al revés
}
\keyword{internal}
