% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{PFDEWD}
\alias{PFDEWD}
\title{59. Precipitation fraction due to extremely wet days
Precipitation at days exceeding the 99percentile divided by total precipitation}
\usage{
PFDEWD(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
PFDEWD
}
\description{
59. Precipitation fraction due to extremely wet days
Precipitation at days exceeding the 99percentile divided by total precipitation
}
\examples{
calculate59(data = pr.value)
}
