% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{PFDVWD}
\alias{PFDVWD}
\title{58. Precipitation fraction due to very wet days
Precipitation at days exceeding the 95percentile divided by total precipitation}
\usage{
PFDVWD(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
PFDVWD
}
\description{
58. Precipitation fraction due to very wet days
Precipitation at days exceeding the 95percentile divided by total precipitation
}
\examples{
calculate58(data = pr.value)
}
