% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{pinna}
\alias{pinna}
\title{101. Pinna Combinative index
 Pinna = 1/2* ((P/(T+10)) + (12*Pd / (Td+10))) where P and T are the multi-annual mean values of precipitation and air temperature, respectively, and P’d and T’d are the mean values of precipitation and air temperature of the driest month, respectively}
\usage{
pinna(pr, taverage, data_names = NULL)
}
\arguments{
\item{pr}{precipitation}

\item{taverage}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
Pinna Combinative index
}
\description{
101. Pinna Combinative index
 Pinna = 1/2* ((P/(T+10)) + (12*Pd / (Td+10))) where P and T are the multi-annual mean values of precipitation and air temperature, respectively, and P’d and T’d are the mean values of precipitation and air temperature of the driest month, respectively
}
\examples{
calculate101(pr = pr.value, taverage = taverage.value)
}
