% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{prcptot}
\alias{prcptot}
\title{Total precipitation wet days}
\usage{
prcptot(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{daily precipitation, mm}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
precipitation, mm
}
\description{
Precipitation amount on days with RR >= 1 mm
}
\examples{
data(data_all)
prcptot(data = data_all$rr)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
