% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{precipitation_concentration_index}
\alias{precipitation_concentration_index}
\title{62. iAnnual/seasonal Precipitation Concentration Index
PCI=100*sum(P)/P^2 (foto en el xlsx)}
\usage{
precipitation_concentration_index(data, data_names = NULL,
  time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Precipitation Concentration Index
}
\description{
62. iAnnual/seasonal Precipitation Concentration Index
PCI=100*sum(P)/P^2 (foto en el xlsx)
}
\examples{
calculate62(data = pr.value)
}
