% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{precipitation_wet_days}
\alias{precipitation_wet_days}
\title{66. Total precipitation in wet days
Precipitation amount on days with RR >= 1 mm in a choosen period (e.g. year)}
\usage{
precipitation_wet_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
precipitation in wet days
}
\description{
66. Total precipitation in wet days
Precipitation amount on days with RR >= 1 mm in a choosen period (e.g. year)
}
\examples{
calculate66(data = pr.value)
}
