% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{quantile_null}
\alias{quantile_null}
\title{quantile null}
\usage{
quantile_null(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}
}
\value{
operation
}
\description{
quantile null
}
\keyword{internal}
