% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{r10mm}
\alias{r10mm}
\title{46. R10mm
Annual count of days when daily precipitation amount >= 10mm}
\usage{
r10mm(data, date, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{date}{date}

\item{time.scale}{month, season or year}
}
\value{
R10mm
}
\description{
46. R10mm
Annual count of days when daily precipitation amount >= 10mm
}
\examples{
calculate46(data = pr.value)
}
