% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{R20mm}
\alias{R20mm}
\title{47. R20mm
Annual count of days when daily precipitation amount >= 20mm}
\usage{
R20mm(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
R20mm
}
\description{
47. R20mm
Annual count of days when daily precipitation amount >= 20mm
}
\examples{
calculate47(data = pr.value)
}
