% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{R95p}
\alias{R95p}
\title{61. R95p
number of days with precipitation above 95 percentile}
\usage{
R95p(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
R95p
}
\description{
61. R95p
number of days with precipitation above 95 percentile
}
\examples{
calculate61(data = pr.value)
}
