% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{R95pTOT}
\alias{R95pTOT}
\title{48. R95pTOT
Annual total PRCP when RR > 95p}
\usage{
R95pTOT(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
R95pTOT
}
\description{
48. R95pTOT
Annual total PRCP when RR > 95p
}
\examples{
calculate48(data = pr.value)
}
