% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{R99pTOT}
\alias{R99pTOT}
\title{49. R99pTOT
Annual total PRCP when RR > 99p}
\usage{
R99pTOT(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
R99pTOT
}
\description{
49. R99pTOT
Annual total PRCP when RR > 99p
}
\examples{
calculate49(data = pr.value)
}
