% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{r_to_in}
\alias{r_to_in}
\title{Transforma radiancia en insolación}
\usage{
r_to_in(radiation, lat, mde)
}
\arguments{
\item{radiation}{radiation}

\item{lat}{lat}

\item{mde}{mde}
}
\value{
insolación en horas
}
\description{
Transforma radiancia en insolación
}
\keyword{internal}
