% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{ref_data}
\alias{ref_data}
\title{Data in reference interval
Does nothing}
\usage{
ref_data(data, date)
}
\arguments{
\item{data}{data}

\item{date}{date}
}
\value{
Data in reference interval
}
\description{
Data in reference interval
Does nothing
}
\examples{
ref_data(data=tmax.value)
}
