% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{RR1}
\alias{RR1}
\title{67. RR1: 
Wet days >= 1 mm (days), ECA&D standard}
\usage{
RR1(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
RR1
}
\description{
67. RR1: 
Wet days >= 1 mm (days), ECA&D standard
}
\examples{
calculate67(data = pr.value)
}
