% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{RR3}
\alias{RR3}
\title{68. RR3: 
Wet days >= 3mm (days), ECA&D standard}
\usage{
RR3(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
RR3
}
\description{
68. RR3: 
Wet days >= 3mm (days), ECA&D standard
}
\examples{
calculate68(data = pr.value)
}
