% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{Rx1day}
\alias{Rx1day}
\title{50. Rx1day
Monthly maximum 1-day precipitation}
\usage{
Rx1day(data, data_names = NULL)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}
}
\value{
Rx1day
}
\description{
50. Rx1day
Monthly maximum 1-day precipitation
}
\examples{
calculate50(data = pr.value)
}
