% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{Rx5day}
\alias{Rx5day}
\title{51. Rx5day
Monthly maximum consecutive 5-day precipitation}
\usage{
Rx5day(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Rx5day
}
\description{
51. Rx5day
Monthly maximum consecutive 5-day precipitation
}
\examples{
calculate51(data = pr.value)
}
