% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{sd0_10}
\alias{sd0_10}
\title{Snow depth 1-10}
\usage{
sd0_10(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
SD0_10
}
\description{
Number of days with snow depth in the range 1-10 cm
}
\examples{
data(data_all)
sd0_10(data = data_all$snowdepththickness)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
