% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{sd10_20}
\alias{sd10_20}
\title{Snow depth 10-20}
\usage{
sd10_20(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snow depth, mm}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
days
}
\description{
The number of days with snow depth of 10-20 cm
}
\examples{
data(data_all)
sd10_20(data = data_all$snowdepth)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
