% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{sdd}
\alias{sdd}
\title{Snow depth}
\usage{
sdd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
snow depth
}
\description{
Mean of daily snow depth
mean of daily snow depth
}
\examples{
data(data_all)
sdd(data = data_all$snowdepththickness)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
