% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{SDII}
\alias{SDII}
\title{52. SDII
Simple precipitation intensity index}
\usage{
SDII(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
SDII
}
\description{
52. SDII
Simple precipitation intensity index
}
\examples{
calculate52(data = pr.value)
}
