% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{seasonals}
\alias{seasonals}
\title{Seasonals}
\usage{
seasonals(time)
}
\arguments{
\item{time}{chron}
}
\value{
seasonals
}
\description{
Seasonals
}
\examples{
seasonals(time=chron(names(tmax.value)))
}
