% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{select_all_time_function}
\alias{select_all_time_function}
\title{Function to select all "time" data}
\usage{
select_all_time_function(time.scale)
}
\arguments{
\item{time.scale}{month, season or year}
}
\value{
function
}
\description{
Function to select all "time" data
}
\keyword{internal}
