% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{select_time_function}
\alias{select_time_function}
\title{Function to select data}
\usage{
select_time_function(time.scale)
}
\arguments{
\item{time.scale}{month, season or year}
}
\value{
function
}
\description{
Function to select data
}
