% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{select_value_for_data}
\alias{select_value_for_data}
\title{Name data station or month}
\usage{
select_value_for_data(data, value, time.scale)
}
\arguments{
\item{data}{data}

\item{value}{value for month, season or year}

\item{time.scale}{month, season or year}
}
\value{
function
}
\description{
Name data station or month
}
\keyword{internal}
