% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{snd}
\alias{snd}
\title{Sunny days}
\usage{
snd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{cloud cover, percentage}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
days
}
\description{
Days with mean cloud cover less than 10%.
}
\examples{
data(data_all)
snd(data = data_all$cloud)
}
