% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{snow_covered_days}
\alias{snow_covered_days}
\title{115. Average amount of snow covered days in period}
\usage{
snow_covered_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
snow covered days
}
\description{
115. Average amount of snow covered days in period
}
\examples{
calculate115(data = snow_depth.value)
}
