% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{snow_depth}
\alias{snow_depth}
\title{107. snow depth
mean of daily snow depth}
\usage{
snow_depth(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snow depth}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
snow depth
}
\description{
107. snow depth
mean of daily snow depth
}
\examples{
calculate107(data = snow_depth.value)
}
