% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{snowfall_sum}
\alias{snowfall_sum}
\title{104. Snowfall sum
The annual, seasonal and monthly sum of snowfall}
\usage{
snowfall_sum(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{snowfall}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
n0to10
}
\description{
104. Snowfall sum
The annual, seasonal and monthly sum of snowfall
}
\examples{
calculate104(data = snow.value)
}
