% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{SPEI}
\alias{SPEI}
\title{125. Standardized Precipitation Evapotranspiration Index (SPEI)
1, 3, 6 and 12 month SPEI}
\usage{
SPEI(eto, pr, data_names = NULL, scale = 3)
}
\arguments{
\item{eto}{et0}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{scale}{scale}
}
\value{
SPEI
}
\description{
125. Standardized Precipitation Evapotranspiration Index (SPEI)
1, 3, 6 and 12 month SPEI
}
\examples{
calculate125(eto = eto.value, pr = pr.value)
}
