% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_ind_MCH.R
\name{spell.length.quantile}
\alias{spell.length.quantile}
\title{Quantile of lenghts of spells}
\usage{
spell.length.quantile(cloud.bin, date.factor, q)
}
\arguments{
\item{cloud.bin}{Time series of binary data (counted are 1s)}

\item{date.factor}{Factor to aggregate by.}

\item{q}{Quantile that should be calculated.}
}
\value{
for each date.factor, the qth quantile of the consecutive 1s
}
\description{
This function calculates the qth quantile of the spell lenghts.
The center of the spell determines to which date.factor the spell is counted to and
theoretically, spells can last longer than the duration of a date.factor.
}
