% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{SPI}
\alias{SPI}
\title{124. Standardized Precipitation Index (SPI)
1, 3, 6 and 12 month SPI}
\usage{
SPI(data, data_names = NULL, scale = 3)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{scale}{scale}
}
\value{
SPI
}
\description{
124. Standardized Precipitation Index (SPI)
1, 3, 6 and 12 month SPI
}
\examples{
calculate124(data = pr.value)
}
