% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{spi3}
\alias{spi3}
\title{Standardized precipitation index 3}
\usage{
spi3(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{daily precipitation, mm}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{further arguments passed to or from other methods}
}
\value{
index value
}
\description{
Standardized precipitation index calculated at 3-month time scale
}
\examples{
data(data_all)
spi3(data = data_all$rr)
}
\references{
McKee, T. B., Doesken, N. J. and Kleist, J.: The relationship of drought frequency and duration to time scales, Eighth Conf. Appl. Climatol., 179–184, 1993.
}
