% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{SS}
\alias{SS}
\title{122. SS: 
Sunshine duration (hours), standard ECA&D}
\usage{
SS(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{sunshine duration}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
SS
}
\description{
122. SS: 
Sunshine duration (hours), standard ECA&D
}
\examples{
calculate122(data = insolation.value)
}
