% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{SSp}
\alias{SSp}
\title{121. SSp:
Sunshine duration fraction with respect to day length (\%), standard ECA&D}
\usage{
SSp(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{sunshine duration}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
SSp
}
\description{
121. SSp:
Sunshine duration fraction with respect to day length (\%), standard ECA&D
}
\examples{
calculate121(data = insolation.value)
}
