% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{stn15}
\alias{stn15}
\title{Sums TN-15}
\usage{
stn15(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{daily minimum temperature, Celsius}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{further arguments passed to or from other methods}
}
\value{
temperature, Celsius
}
\description{
Sum of degree days when TN <= -15 Celsius recorded in December-February interval
}
\examples{
data(data_all)
stn15(data = data_all$tn)
}
