% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{sumf}
\alias{sumf}
\title{sum with na.rm TRUE}
\usage{
sumf(..., na.rm = TRUE)
}
\arguments{
\item{...}{...}

\item{na.rm}{na.rm}
}
\value{
sum
}
\description{
sum with na.rm TRUE
}
\keyword{internal}
