% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{summer_days}
\alias{summer_days}
\title{19. Summer days
Number of days with maximum temperature >25ºC per year.}
\usage{
summer_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Summer days
}
\description{
19. Summer days
Number of days with maximum temperature >25ºC per year.
}
\examples{
calculate19(data=tmax.value)
}
