% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{sunny_days}
\alias{sunny_days}
\title{118. sunny days
days with mean cloud cover less than 10\%.}
\usage{
sunny_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{cloud cover}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
sunny days
}
\description{
118. sunny days
days with mean cloud cover less than 10\%.
}
\examples{
calculate118(data = cloud_cover.value)
}
