% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{sunshine_duration}
\alias{sunshine_duration}
\title{117. Sum of sunshine duration}
\usage{
sunshine_duration(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{sunshine duration}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
sunshine duration
}
\description{
117. Sum of sunshine duration
}
\examples{
calculate117(data = insolation.value)
}
