% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ta_o}
\alias{ta_o}
\title{Growing season (Apr-Oct)}
\usage{
ta_o(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
Growing season temperature 1
}
\description{
Growing season (april to october) mean TG
}
\examples{
data(data_all)
ta_o(data=data_all$tg)
}
