% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{TCI}
\alias{TCI}
\title{132. TCI: Tourism Climatic Index
Standard index computed by ECA&D; Described at Miezkowski (1985), conceptual formula: TCI = 4cid + cia + 2R + 2S + W, where CId is a daytime comfort index, CIa a daily comfort index, R is cumulated rainfall, S the daily sunshine hours and W wind speed}
\usage{
TCI(data, sunshine, w, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{sunshine}{net radiation}

\item{w}{average wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
TCI
}
\description{
132. TCI: Tourism Climatic Index
Standard index computed by ECA&D; Described at Miezkowski (1985), conceptual formula: TCI = 4cid + cia + 2R + 2S + W, where CId is a daytime comfort index, CIa a daily comfort index, R is cumulated rainfall, S the daily sunshine hours and W wind speed
}
\examples{
calculate132(data=pr.value, sunshine=radiation.value, w=w.value)
}
