% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{TCI60}
\alias{TCI60}
\title{133. TCI>60: 
Number of days TCI>60 , standard ECA&D}
\usage{
TCI60(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{Tourism Climatic Index}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
TCI60
}
\description{
133. TCI>60: 
Number of days TCI>60 , standard ECA&D
}
\examples{
calculate133(data = tci.value)
}
