% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{TCI80}
\alias{TCI80}
\title{134. TCI>80:
Number of days TCI>80, standard ECA&D}
\usage{
TCI80(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{Tourism Climatic Index}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
TCI80
}
\description{
134. TCI>80:
Number of days TCI>80, standard ECA&D
}
\examples{
calculate134(data = tci.value)
}
