% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{td_to_rh}
\alias{td_to_rh}
\title{Dew point to relative humidity}
\usage{
td_to_rh(tmax, tmin, td)
}
\arguments{
\item{tmax}{maximum temperature}

\item{tmin}{minimum temperature}

\item{td}{dew point}
}
\value{
rh
}
\description{
Dew point to relative humidity
}
\examples{
td_to_rh(tmax = tmax.value, tmin = tmin.value, td = dew_point.value)
}
