% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{td_to_vapor}
\alias{td_to_vapor}
\title{Dew point to water vapour pressure}
\usage{
td_to_vapor(td)
}
\arguments{
\item{td}{dew point}
}
\value{
vapor
}
\description{
Dew point to water vapour pressure
}
\examples{
td_to_vapor(td = dew_point.value)
}
