% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_average_1.2}
\alias{temperature_average_1.2}
\title{38.2. Temperature average 1
Average number of days whith maximum temperature >= 32ºC in June, July, August and from June to August}
\usage{
temperature_average_1.2(data, data_names = NULL)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 1
}
\description{
38.2. Temperature average 1
Average number of days whith maximum temperature >= 32ºC in June, July, August and from June to August
}
\examples{
calculate38.2(data = tmax.value)
}
