% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums}
\alias{temperature_sums}
\title{21. Temperature sums
Sum of Tmax days >17ºC–days Tmax <17ºC ºC}
\usage{
temperature_sums(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Temperature sums
}
\description{
21. Temperature sums
Sum of Tmax days >17ºC–days Tmax <17ºC ºC
}
\examples{
calculate21(data=tmax.value)
}
