% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_1.1}
\alias{temperature_sums_1.1}
\title{38.1 Temperature sums 1
Sums of maximum temperatures >= 32ºC in June, July, August and from June to August}
\usage{
temperature_sums_1.1(data, data_names = NULL)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 1
}
\description{
38.1 Temperature sums 1
Sums of maximum temperatures >= 32ºC in June, July, August and from June to August
}
\examples{
calculate38.1(data = tmax.value)
}
