% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_1.1_time}
\alias{temperature_sums_1.1_time}
\title{38.1 Temperature sums 1
Sums of maximum temperatures >= 32ºC in June, July, August or from June to August}
\usage{
temperature_sums_1.1_time(data, data_names = NULL, selectTime)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{selectTime}{Jun, Jul, Aug or from Jun to Aug (summer)}
}
\value{
temperature sums 1
}
\description{
38.1 Temperature sums 1
Sums of maximum temperatures >= 32ºC in June, July, August or from June to August
}
\examples{
calculate38.1_time(data = tmax.value, selectTime=JUN)
}
