% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_2}
\alias{temperature_sums_2}
\title{39. Temperature sums 2
Sums of minimum air temperatures <= -15ºC recorded in December-February interval}
\usage{
temperature_sums_2(data, data_names = NULL)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 2
}
\description{
39. Temperature sums 2
Sums of minimum air temperatures <= -15ºC recorded in December-February interval
}
\examples{
calculate39(data = tmin.value)
}
