% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_3}
\alias{temperature_sums_3}
\title{40. Temperature sums 3
Sums of minimum air temperatures <=-10ºC recorded in December-February interval}
\usage{
temperature_sums_3(data, data_names = NULL)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 3
}
\description{
40. Temperature sums 3
Sums of minimum air temperatures <=-10ºC recorded in December-February interval
}
\examples{
calculate40(data = tmin.value)
}
