% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_4}
\alias{temperature_sums_4}
\title{41. Temperature sums 4
The average daily amount of negative air temperature (sum of Tmed <= 0ºC/ cold units) recorded in the November to March period}
\usage{
temperature_sums_4(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 4
}
\description{
41. Temperature sums 4
The average daily amount of negative air temperature (sum of Tmed <= 0ºC/ cold units) recorded in the November to March period
}
\examples{
calculate41(data = taverage.value)
}
