% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_5}
\alias{temperature_sums_5}
\title{42. Temperature sums 5
Sums of positive average temperatures calculated for the 1st of February to the 10th April interval}
\usage{
temperature_sums_5(data, data_names = NULL)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 5
}
\description{
42. Temperature sums 5
Sums of positive average temperatures calculated for the 1st of February to the 10th April interval
}
\examples{
calculate42(data = taverage.value)
}
