% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{temperature_sums_6}
\alias{temperature_sums_6}
\title{43. Temperature sums 6
Number of consecutive days whith maximum temperature >= 32ºC from June to August}
\usage{
temperature_sums_6(data, data_names = NULL)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}
}
\value{
temperature sums 6
}
\description{
43. Temperature sums 6
Number of consecutive days whith maximum temperature >= 32ºC from June to August
}
\examples{
calculate43(data = tmax.value)
}
