% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{tm_s}
\alias{tm_s}
\title{Growing season(May-Sep)}
\usage{
tm_s(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
Growing season temperature 2
}
\description{
Growing season (may to september) mean TG
}
\examples{
data(data_all)
tm_s(data=data_all$tg)
}
