% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{tn90p}
\alias{tn90p}
\title{Warm nights}
\usage{
tn90p(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{daily minimum temperature, Celsius}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
days
}
\description{
Percentages of days with TN higher than the 90th percentile.  The 90th percentile is computed based on the time scale selected (month, season or year) not daily.
}
\examples{
data(data_all)
tn90p(data=data_all$tn)
}
