% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{TNn}
\alias{TNn}
\title{11. Coldest night (TNn)
Lowest daily minimum temperature at the annual and monthly scales}
\usage{
TNn(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
TNn
}
\description{
11. Coldest night (TNn)
Lowest daily minimum temperature at the annual and monthly scales
}
\examples{
calculate11(data=tmin.value)
}
