% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{total_precipitation}
\alias{total_precipitation}
\title{44. Total precipitation}
\usage{
total_precipitation(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
total precipitation
}
\description{
44. Total precipitation
}
\examples{
calculate44(data = pr.value, time.scale=YEAR)
}
