% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{tropical_nights}
\alias{tropical_nights}
\title{22. Tropical nights
Number of days with minimum temperature >20ºC per year.}
\usage{
tropical_nights(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Tropical nights
}
\description{
22. Tropical nights
Number of days with minimum temperature >20ºC per year.
}
\examples{
calculate22(data=tmin.value)
}
