% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{TXn}
\alias{TXn}
\title{29. Warmest day (TXn)
Highest daily maximum temperature at the annual and monthly scales}
\usage{
TXn(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
TXn
}
\description{
29. Warmest day (TXn)
Highest daily maximum temperature at the annual and monthly scales
}
\examples{
calculate29(data=tmax.value)
}
