% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{uai}
\alias{uai}
\title{UNEP Aridity Index}
\usage{
uai(eto, pr, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{eto}{et0}

\item{pr}{precipitation}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
iUNEP
}
\description{
P/Eto
}
\examples{
data(data_all)
uai(eto = data_all$evapotranspiration, pr = data_all$rr)
}
\references{
Huiping Huang, Yuping Han, Mingming Cao, Jinxi Song, and Heng Xiao Spatial-Temporal Variation of Aridity Index of China during 1960–2013. Advances in Meteorology, vol. 2016, Article ID 1536135, 10 pages, 2016. \url{https://doi.org/10.1155/2016/1536135}
}
