% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{vDTR}
\alias{vDTR}
\title{13. vDTR
Mean absolute day-to-day difference in DTR (annual and monthly)}
\usage{
vDTR(max, min, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{max}{maximum temperature}

\item{min}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
vDTR
}
\description{
13. vDTR
Mean absolute day-to-day difference in DTR (annual and monthly)
}
\examples{
calculate13(max=tmax.value, min=tmin.value)
}
