% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{very_cold_days}
\alias{very_cold_days}
\title{24. Very cold days
Number of days with minimum temperature <1st percentile per year.}
\usage{
very_cold_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Very cold days
}
\description{
24. Very cold days
Number of days with minimum temperature <1st percentile per year.
}
\examples{
calculate24(data=tmin.value)
}
