% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{very_warm_days}
\alias{very_warm_days}
\title{25. Very warm days
Number of days with maximum temperature >99th percentile per year.}
\usage{
very_warm_days(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Very warm days
}
\description{
25. Very warm days
Number of days with maximum temperature >99th percentile per year.
}
\examples{
calculate25(data=tmax.value)
}
