% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{vwd}
\alias{vwd}
\title{Very warm days}
\usage{
vwd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Very warm days
}
\description{
Days with TX >99th percentile per year.
}
\examples{
data(data_all)
vwd(data=data_all$tx)
}
