% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{warm_nights}
\alias{warm_nights}
\title{27. Warm nights
Percentages of days with minimum temperatures higher than the 90th percentile.}
\usage{
warm_nights(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Warm nights
}
\description{
27. Warm nights
Percentages of days with minimum temperatures higher than the 90th percentile.
}
\examples{
calculate27(data=tmin.value)
}
