% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{warm_spell_duration_index}
\alias{warm_spell_duration_index}
\title{28. Warm spell duration index
Annual count of days with at least 6 consecutive days when TX > 90th percentile}
\usage{
warm_spell_duration_index(data, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
Warm spell duration index
}
\description{
28. Warm spell duration index
Annual count of days with at least 6 consecutive days when TX > 90th percentile
}
\examples{
calculate28(data=tmax.value)
}
