% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{WCI}
\alias{WCI}
\title{87. Wind Chill Index: temperature + wind 
Osczevski, R. & Bleustein, M. 2005, Bull. Amer. Meteor. Soc., 86, 1453, doi:10.1175/BAMS-86-10-1453 https://journals.ametsoc.org/doi/abs/10.1175/BAMS-86-10-1453
Ta = air temperature in fahrenheit; in ºC ; v = wind speed in km/h
http://www.calculator.net/wind-chill-calculator.html
WCI = 35.74 + 0.6215*T - 35.75*V0.16 + 0.4275*T*V0.16
Old WCI = 0.081*(3.71*V1/2 + 5.81 - 0.25*V) × (T - 91.4) + 91.4}
\usage{
WCI(taverage, w, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{taverage}{medium temperature}

\item{w}{average wind}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
WCI
}
\description{
87. Wind Chill Index: temperature + wind 
Osczevski, R. & Bleustein, M. 2005, Bull. Amer. Meteor. Soc., 86, 1453, doi:10.1175/BAMS-86-10-1453 https://journals.ametsoc.org/doi/abs/10.1175/BAMS-86-10-1453
Ta = air temperature in fahrenheit; in ºC ; v = wind speed in km/h
http://www.calculator.net/wind-chill-calculator.html
WCI = 35.74 + 0.6215*T - 35.75*V0.16 + 0.4275*T*V0.16
Old WCI = 0.081*(3.71*V1/2 + 5.81 - 0.25*V) × (T - 91.4) + 91.4
}
\examples{
calculate87(taverage = taverage.value, w = w.value)
}
