% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{wd}
\alias{wd}
\title{Warm days}
\usage{
wd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Warm days
}
\description{
Total numbers of days with TX higher than the 90th percentile.
}
\examples{
data(data_all)
wd(data=data_all$tx)
}
