% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/custom_functions.R
\name{whichf}
\alias{whichf}
\title{which with arr.ind TRUE}
\usage{
whichf(x, arr.ind = TRUE, useNames = TRUE)
}
\arguments{
\item{x}{x}

\item{arr.ind}{na.rm}

\item{useNames}{na.rm}
}
\value{
which
}
\description{
which with arr.ind TRUE
}
