% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{winkler_index}
\alias{winkler_index}
\title{36. Winkler index
Sum of degree days over 10°C from April 1 until October 31 = Sum max [(avg. daily temp. – 10), 0]}
\usage{
winkler_index(data, data_names = NULL)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}
}
\value{
Winkler index
}
\description{
36. Winkler index
Sum of degree days over 10°C from April 1 until October 31 = Sum max [(avg. daily temp. – 10), 0]
}
\examples{
calculate36(data = tmax.value)
}
