% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{winter_severity_index}
\alias{winter_severity_index}
\title{37. Winter Severity index
Mean temperature of the coldest month of the year}
\usage{
winter_severity_index(data, data_names = NULL)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}
}
\value{
Winter Severity index
}
\description{
37. Winter Severity index
Mean temperature of the coldest month of the year
}
\examples{
calculate37(data = tmax.value)
}
