% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{wki}
\alias{wki}
\title{Winkler index}
\usage{
wki(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
Winkler index
}
\description{
Sum of degree days over 10 celsius of TG from April 1 until October 31
}
\examples{
data(data_all)
wki(data = data_all$tg)
}
\references{
Winkler, A.J., J.A. Cook, W.M. Kliewer, and L.A. Lider. 1974. General Viticulture. 4th ed. University of California Press, Berkeley.
}
