% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{wn}
\alias{wn}
\title{Warm nights}
\usage{
wn(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Warm nights
}
\description{
Percentages of days with TN higher than the 90th percentile.
}
\examples{
data(data_all)
wn(data=data_all$tn)
}
