% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{wsd}
\alias{wsd}
\title{Warm spell duration}
\usage{
wsd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Warm spell duration index
}
\description{
Count of days with at least 6 consecutive days when TX > 90th percentile.
}
\examples{
data(data_all)
wsd(data=data_all$tx)
}
