% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{zcd}
\alias{zcd}
\title{Zero crossing days}
\usage{
zcd(tmax, tmin, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{tmax}{maximum temperature}

\item{tmin}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
zero crossing days
}
\description{
Number of days with TX > 0 Celsius and TN < 0 Celsius.
}
\examples{
data(data_all)
zcd(tmax=data_all$tx, tmin=data_all$tn)
}
