% Generated by roxygen2: do not edit by hand
% Please edit documentation in paquete_sequia/indecis.R
\name{zero_crossing}
\alias{zero_crossing}
\title{30. zero crossing days
Number of days with Tmax > 0 ºC and Tmin < 0 ºC.}
\usage{
zero_crossing(max, min, data_names = NULL, time.scale = YEAR)
}
\arguments{
\item{max}{maximum temperature}

\item{min}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}
}
\value{
zero crossing days
}
\description{
30. zero crossing days
Number of days with Tmax > 0 ºC and Tmin < 0 ºC.
}
\examples{
calculate30(max=tmax.value, min=tmin.value)
}
