% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{average_temp}
\alias{average_temp}
\title{Average temperature}
\usage{
average_temp(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum, minimum or medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
average temperature
}
\description{
Average temperature
}
\keyword{internal}
