% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{bio17}
\alias{bio17}
\title{Precipitation of Driest Quarter}
\usage{
bio17(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
BIO17
}
\description{
Precipitation of the driest quarter of the year
}
\examples{
data(data_all)
bio17(data = data_all$rr)
}
\references{
Hijmans RJ, Cameron SE, Parra JL, Jones PG, Jarvis A (2005) Very high resolution interpolated climate surfaces for global land areas. Int J Climatol 25:1965–1978. doi: 10.1002/joc.1276. \url{http://www.worldclim.org/bioclim}
}
