% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{bio19}
\alias{bio19}
\title{Precipitation coldest quarter}
\usage{
bio19(pr, taverage, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{pr}{daily precipitation, mm}

\item{taverage}{daily mean temperature, Celsius}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{further arguments passed to or from other methods}
}
\value{
precipitation, mm
}
\description{
Precipitation of the coldest quarter of the year
}
\examples{
data(data_all)
bio19(pr=data_all$rr, taverage=data_all$tg)
}
\references{
Hijmans RJ, Cameron SE, Parra JL, Jones PG, Jarvis A (2005) Very high resolution interpolated climate surfaces for global land areas. Int J Climatol 25:1965–1978. \doi{10.1002/joc.1276}. \url{https://web.archive.org/web/20190714191708/https://www.worldclim.org/bioclim}
}
