% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{bio5}
\alias{bio5}
\title{TX warmest month}
\usage{
bio5(data, tmax, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{mean temperature}

\item{tmax}{maximum temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
BIO5
}
\description{
TX of the warmest month of the year
}
\examples{
data(data_all)
bio5(data = data_all$tg, tmax = data_all$tx)
}
\references{
Hijmans RJ, Cameron SE, Parra JL, Jones PG, Jarvis A (2005) Very high resolution interpolated climate surfaces for global land areas. Int J Climatol 25:1965–1978. doi: 10.1002/joc.1276. \url{http://www.worldclim.org/bioclim}
}
