% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{bio7}
\alias{bio7}
\title{Temperature Annual Range}
\usage{
bio7(data, tmin, tmax, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{medium temperature}

\item{tmin}{min temperature}

\item{tmax}{max temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
BIO7
}
\description{
TX of the warmest month minus TN of coldest month
}
\examples{
data(data_all)
bio7(data = data_all$tg, tmin = data_all$tn, tmax = data_all$tx)
}
\references{
Hijmans RJ, Cameron SE, Parra JL, Jones PG, Jarvis A (2005) Very high resolution interpolated climate surfaces for global land areas. Int J Climatol 25:1965–1978. doi: 10.1002/joc.1276. \url{http://www.worldclim.org/bioclim}
}
