% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{byMonths_chron}
\alias{byMonths_chron}
\title{Data with months and years in names}
\usage{
byMonths_chron(data)
}
\arguments{
\item{data}{data}
}
\value{
dates
}
\description{
Data with months and years in names
}
\keyword{internal}
