% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{calc_eto}
\alias{calc_eto}
\title{Et0}
\usage{
calc_eto(
  tmin,
  tmax,
  toa,
  w,
  mde,
  lat,
  tdew,
  radiation = NA,
  insolation = NA,
  rh = NA,
  na.rm = FALSE
)
}
\arguments{
\item{tmin}{daily minimum temperature, Celsius, Celsius}

\item{tmax}{maximum temperature, Celsius}

\item{toa}{radiation toa, J/m2}

\item{w}{average wind, m/s at 10m}

\item{mde}{mde}

\item{lat}{latitude}

\item{tdew}{dew point, Celsius}

\item{radiation}{radiation, J m-2}

\item{insolation}{insolation, hours}

\item{rh}{relative humidity, percentage}

\item{na.rm}{na.rm}
}
\value{
et0
}
\description{
Et0
}
\keyword{internal}
