% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{calcf_data}
\alias{calcf_data}
\title{Operation de data, agrupando los datos por valores de names}
\usage{
calcf_data(
  data,
  date,
  time.scale,
  extract_names = select_time_function,
  data_names,
  operation,
  ...
)
}
\arguments{
\item{data}{data}

\item{date}{date}

\item{time.scale}{month, season, year or hydrological_years}

\item{extract_names}{Operation to split data}

\item{data_names}{names of each period of time}

\item{operation}{Main operation}

\item{...}{...}
}
\value{
result operation
}
\description{
Operation de data, agrupando los datos por valores de names
}
\keyword{internal}
