% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{calcf_data_}
\alias{calcf_data_}
\title{Operation de data, agrupando los datos por valores de names}
\usage{
calcf_data_(data_names, data, operation, time.scale, ...)
}
\arguments{
\item{data_names}{names}

\item{data}{data}

\item{operation}{operation}

\item{time.scale}{month, season, year or hydrological_years}

\item{...}{...}
}
\value{
operation
}
\description{
Operation de data, agrupando los datos por valores de names
}
\keyword{internal}
