% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis.R
\name{calculate_all_scales}
\alias{calculate_all_scales}
\title{Calculate all indexes for all time scales}
\usage{
calculate_all_scales(data, lat = NULL)
}
\arguments{
\item{data}{data list}

\item{lat}{latitude, degree}
}
\value{
all indexes
}
\description{
Calculate all indexes for a point and all time scales
}
