% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{cmd}
\alias{cmd}
\title{Climatic moisture deficit}
\usage{
cmd(
  eto,
  evapotranspiration,
  data_names = NULL,
  time.scale = YEAR,
  na.rm = FALSE
)
}
\arguments{
\item{eto}{eto, mm}

\item{evapotranspiration}{evapotranspiration, mm}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
index value
}
\description{
ETo - evapotranspiration
}
\examples{
data(data_all)
cmd(eto = data_all$eto, evapotranspiration = data_all$evaporation)
}
\references{
Parks, S. A., Parisien, M. , Miller, C. , Holsinger, L. M. and Baggett, L. S. (2018), Fine-scale spatial climate variation and drought mediate the likelihood of reburning. Ecol Appl, 28: 573-586. \doi{10.1002/eap.1671}
}
