% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{d32}
\alias{d32}
\title{Days TX32}
\usage{
d32(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{maximum temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
temperature sums 1
}
\description{
Number of days whith TX >= 32 Celsius on the interval June-August.
}
\examples{
data(data_all)
d32(data = data_all$tx)
}
