% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Datasets}
\alias{Datasets}
\alias{data_all}
\title{data_all}
\format{
An object of class \code{list} of length 22.
}
\usage{
data(data_all)
}
\description{
See wichita
}
\details{
See description.
}
\keyword{datasets}
