% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{fg}
\alias{fg}
\title{Mean of daily mean wind strength}
\usage{
fg(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{average wind, m/s}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
wind, m/s
}
\description{
Mean of daily FG
}
\examples{
data(data_all)
fg(data = data_all$wind)
}
\references{
European Climate Assessment & Dataset. Indices dictionary. \url{https://www.ecad.eu//indicesextremes/indicesdictionary.php}
}
