% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{gd4}
\alias{gd4}
\title{Growing degree days}
\usage{
gd4(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
GD4
}
\description{
Sum of degree days of TG over 4 Celsius (the daily mean temperature is less than 4 celsius, it is set equal to 4 celsius)
}
\examples{
data(data_all)
gd4(data=data_all$tg)
}
\references{
McMaster, G. S., & Wilhelm, W. W. (1997). Growing degree-days: One equation, two interpretations. Agricultural and Forest Meteorology, 87(4), 291-300
}
