% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{gsl}
\alias{gsl}
\title{Growing season length}
\usage{
gsl(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{mean temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
growing season length
}
\description{
Annual count of days between the first span of at least 6 days with TG > 5 Celsius and first span after 1 July of 6 days with TG < 5 Celsius.
}
\examples{
data(data_all)
gsl(data=data_all$tg)
}
\references{
Klein Tank AMG, Zwiers FW, Zhang X. 2009. Guidelines on analysis of extremes in a changing climate in support of informed decisions for adaptation, climate data and monitoring WCDMP-No 72, WMO-TD No 1500, p 5.
}
