% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{gsr}
\alias{gsr}
\title{Growing season precipitation}
\usage{
gsr(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{precipitation}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
GSP
}
\description{
Growing season (april to october) total precipitation
}
\examples{
data(data_all)
gsr(data = data_all$rr)
}
