% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{maximum_temp}
\alias{maximum_temp}
\title{Maximum temperature}
\usage{
maximum_temp(data, data_names = NULL, time.scale = YEAR,
  na.rm = FALSE)
}
\arguments{
\item{data}{maximum, minimum or medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
maximum temperature
}
\description{
Maximum temperature
}
\keyword{internal}
