% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{minimum_temp}
\alias{minimum_temp}
\title{Minimum temperature}
\usage{
minimum_temp(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{maximum, minimum or medium temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
minimum temperature
}
\description{
Minimum temperature
}
\keyword{internal}
