% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{months_years}
\alias{months_years}
\title{Months by years}
\usage{
months_years(time)
}
\arguments{
\item{time}{chron}
}
\value{
months by years
}
\description{
Months by years
}
\keyword{internal}
