% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ptg}
\alias{ptg}
\title{Sums positive}
\usage{
ptg(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{medium temperature}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{...}
}
\value{
temperature sums 5
}
\description{
Sums of positive TG calculated for the 1st of February to the 10th April interval
}
\examples{
data(data_all)
ptg(data = data_all$tg)
}
