% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{seasonals_years}
\alias{seasonals_years}
\title{Seasonals by years}
\usage{
seasonals_years(time)
}
\arguments{
\item{time}{chron}
}
\value{
seasonals by years
}
\description{
Seasonals by years
}
\keyword{internal}
