% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{ss}
\alias{ss}
\title{Snowfall sum}
\usage{
ss(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{snowfall, mm of water equivalent}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
snow, mm
}
\description{
Sum of snowfall
}
\examples{
data(data_all)
ss(data = data_all$snowfallmm)
}
