% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{td_to_rh}
\alias{td_to_rh}
\title{Dew point to relative humidity}
\usage{
td_to_rh(tmax, tmin, td)
}
\arguments{
\item{tmax}{maximum temperature}

\item{tmin}{daily minimum temperature, Celsius}

\item{td}{dew point}
}
\value{
rh
}
\description{
Dew point to relative humidity
}
\keyword{internal}
