% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices_functions.R
\name{td_to_vapor}
\alias{td_to_vapor}
\title{Dew point to water vapour pressure}
\usage{
td_to_vapor(td)
}
\arguments{
\item{td}{dew point}
}
\value{
vapor
}
\description{
Dew point to water vapour pressure
}
\keyword{internal}
