% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{tm_s}
\alias{tm_s}
\title{Growing season(May-Sep)}
\usage{
tm_s(data, data_names = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{data}{daily mean temperature, Celsius}

\item{data_names}{names of each period of time}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}

\item{...}{further arguments passed to or from other methods}
}
\value{
temperature, Celsius
}
\description{
Growing season (may to september) mean TG
}
\examples{
data(data_all)
tm_s(data=data_all$tg)
}
