% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indecis_indices.R
\name{vcd}
\alias{vcd}
\title{Very cold days}
\usage{
vcd(data, data_names = NULL, time.scale = YEAR, na.rm = FALSE)
}
\arguments{
\item{data}{minimum temperature}

\item{data_names}{names of each period of time}

\item{time.scale}{month, season or year}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
Very cold days
}
\description{
Days with TN <1st percentile.
}
\examples{
data(data_all)
vcd(data=data_all$tn)
}
